% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{any_duplicated}
\alias{any_duplicated}
\title{wrapper arround `any(duplicated(.x))`}
\usage{
any_duplicated(.x)
}
\arguments{
\item{.x}{vector of e.g. characters}
}
\description{
wrapper arround `any(duplicated(.x))`
}
