% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{any_missing}
\alias{any_missing}
\title{wrapper arround `stop_if(any(is.na(.x)))`. If `.empty_string_as_missing` is set to TRUE, it also treats
   empty strings as missing values. Otherwise, is just checks for NA values.}
\usage{
any_missing(.x, .empty_string_as_missing = FALSE)
}
\arguments{
\item{.x}{vector of e.g. characters}

\item{.empty_string_as_missing}{if TRUE, empty strings are treated as missing. If TRUE, .x must be a character or factor vector}
}
\description{
wrapper arround `stop_if(any(is.na(.x)))`. If `.empty_string_as_missing` is set to TRUE, it also treats
   empty strings as missing values. Otherwise, is just checks for NA values.
}
