% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{append_list}
\alias{append_list}
\title{`append` appears not to work with dates any longer... in 4.0.2.. perhaps they will fix
e.g.:
    selections <- list()
    append(selections, Sys.Date())  # do not know how to convert 'e' to class “Date”
    selections <- list(Sys.Date())
    append(selections, Sys.Date())  # adds date as numeric.. this use to work}
\usage{
append_list(.list_a, .values)
}
\arguments{
\item{.list_a}{list to append to}

\item{.values}{any value or set of values to append to the list}
}
\description{
`append` appears not to work with dates any longer... in 4.0.2.. perhaps they will fix
e.g.:
    selections <- list()
    append(selections, Sys.Date())  # do not know how to convert 'e' to class “Date”
    selections <- list(Sys.Date())
    append(selections, Sys.Date())  # adds date as numeric.. this use to work
}
