% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_get_param}
\alias{aws_get_param}
\title{AWS tools}
\usage{
aws_get_param(parameter, region = "eu-west-2", profile = "")
}
\arguments{
\item{parameter}{Character string AWS parameter name}

\item{region}{Character string AWS region}

\item{profile}{Character string name of AWS CLI profile if required}
}
\value{
A character string
}
\description{
Retrieve an AWS parameter via SSM
}
\examples{
aws_get_param("param-name")
}
