% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{bundle_case}
\alias{bundle_case}
\title{Bundle case variants}
\usage{
bundle_case(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector matching the length of the input
}
\description{
Bundle all case variations into the most common variant (e.g. hashtags, social media profile names)
}
\examples{
bundle_case(c('london','London','London','Paris','PARIS','paris','Paris'))
}
