% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{centeredColumns}
\alias{centeredColumns}
\title{A centered copy of \code{X}.}
\usage{
centeredColumns(X)
}
\arguments{
\item{X}{a matrix whose columns hold the samples and whose rows hold the
variables}
}
\value{
a matrix with zero-mean samples (one sample per column)
}
\description{
A centered copy of \code{X}.
}

