% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{centeredRows}
\alias{centeredRows}
\title{A centered copy of \code{X}.}
\usage{
centeredRows(X)
}
\arguments{
\item{X}{a matrix whose rows hold the samples and whose columns hold the
variables.}
}
\value{
a matrix with zero-mean samples (one sample per row) and attribute
 `mu`
}
\description{
A centered copy of \code{X}.
}

