% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/cf.R
\name{cf}
\alias{cf}
\title{Copy a file}
\usage{
cf(from, to, overwrite = FALSE, recursive = FALSE, copy.mode = TRUE)
}
\arguments{
\item{overwrite}{logical; should existing destination files be overwritten? (not yet implemented for remote host)}

\item{recursive}{logical.  If 'to' is a directory, should directories in 'from' be copied (and their contents)? (not yet implemented for remote host)}

\item{copy.mode}{logical: should file permission bits be copied where possible? (not yet implemented for remote host)}

\item{from.to}{character vectors, containing file names or paths.}
}
\description{
This function is an extension of file.copy that can also copy to/from on a remote host.
}
\examples{
## On remote host,
lf("host:~/user/data/data.hdf","~/user/data/")
}
\keyword{copy,}
\keyword{remote}

