% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_database.R
\name{checkfolder}
\alias{checkfolder}
\title{Quickly perform simple checks on an EnvloggerViewer folder}
\usage{
checkfolder(PATH, nchar = 5, min_size = 700, plot = TRUE)
}
\arguments{
\item{nchar}{Integer indicating the number of characters to be used to determine the shore name}

\item{min_size}{Integer indicating the file size below which report files are rejected into a "check" folder (defaults to 700, which is just a few bytes greater than the size of a report with header but no data)}

\item{plot}{Logical, defaults to TRUE; set to FALSE to inhibit the diagnostic plot}
}
\value{
Prints diagnostic stats and returns invisibly a list with $dat and $pbox with $dat being a tibble with stats about each valid report found in the folder (i.e., reports with more than 2 lines of data and custom names that aren't empty) and $box being a boxplot of the temperatures recorded by each id. To plot the plot, simply run .$box.
}
\description{
Given a path to a folder from EnvloggerViewer, read data and present several pieces of information that help assessing if there are any issues with the files and data that must be addressed.
}
\examples{
PATH <- dirname(rtools_example("2023-08-31_ukhol")[1])
out <- checkfolder(PATH)
out$dat # tibble with extra details
out$box # boxplot showing the spread of temperatures recorded by each id
}
\seealso{
\code{\link[=read_env_log]{read_env_log()}}
}
