% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{clamp}
\alias{clamp}
\title{Clamp the values of \code{x} to a given range.}
\usage{
clamp(x, lower = -1, upper = +1)
}
\arguments{
\item{x}{input vector}

\item{lower}{the lower bound}

\item{upper}{the upper bound}
}
\value{
a vector of the values of \code{x} clamped to [lower,upper]
}
\description{
Clamp the values of \code{x} to a given range.
}

