% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{coastline_iberia}
\alias{coastline_iberia}
\title{Determine a coastline and find its linear sequence}
\usage{
coastline_iberia(mask, sea_pixels = TRUE, water = 0.1, PLOT = TRUE)
}
\arguments{
\item{mask}{A raster sea-land mask, with sea pixels = 0 and land pixels = 1 (fractional values for transition pixels are allowed but not required).}

\item{sea_pixels}{Logical indicating if the shoreline is composed of the sea pixels closest to land (\code{sea_pixels = TRUE}) or land pixels closest to sea (\code{sea_pixels = FALSE}).}

\item{water}{Numerical from (0,1] (fractional values allowed) used as cutpoint for sea (\verb{<= water}) and land (\verb{> water}) pixels.}

\item{PLOT}{Locical; if \code{TRUE}, the progress of the computation is shown as a plot that is continuously updated}
}
\value{
A list with two elements: \verb{$map} is a raster with the same extent as the input mask, and with all cells set to NA except those corresponding to the coastline, which are numbered sequentially; \verb{$pix} is a tibble with all the coastline cells identified and the following columns: \code{i} = sequential index, \code{x} = lon, \code{y} = lat, and \code{cell} = cell number.
}
\description{
Given a raster mask, determine the coastline and number each of the coastal pixels sequentially (currently only supports the France-Iberia-Morocco region)
}
\examples{
# Load a raster mask
mask <- terra::rast(system.file("extdata/ERA5_mask.nc", package = "rtools"))

# Determine the coastline and find its N-to-S linear sequence
coastline_iberia(mask)
}
