% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{compose}
\alias{compose}
\title{The composition of the given functions \code{...}.}
\usage{
compose(...)
}
\arguments{
\item{...}{some function objects}
}
\value{
a function object that represents the composition of the given
 functions, e.g. \code{f(g(...))} for given \code{f} and \code{g}
}
\description{
The composition of the given functions \code{...}.
}
\examples{
 g <- function(x) 2*x
 h <- function(y) y+3
 f <- compose(g, h)
 f(5) == 16
}

