% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_csv.R
\name{concat_csv}
\alias{concat_csv}
\title{concat_csv Function}
\usage{
concat_csv(csv_list)
}
\arguments{
\item{list}{of paths to the csv files}
}
\description{
This function concatenates multiple csv files (of the same structure) into a master file
}
\keyword{concatenate}
