% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{confusionMatrixScores}
\alias{confusionMatrixScores}
\title{Accuracy, precisions and recalls for a given covariance matrix.}
\usage{
confusionMatrixScores(X)
}
\arguments{
\item{X}{a NxN matrix where rows represent the actual classes and columns
represent the predicted classes}
}
\value{
a list consisting of
 \item{accuracy}{the overall classification accuracy}
 \item{precision}{a vector of the classification precision for each class}
 \item{recall}{a vector of the classification recall for each class}
}
\description{
Accuracy, precisions and recalls for a given covariance matrix.
}

