% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_codes.R
\name{country_to_iso3}
\alias{country_to_iso3}
\title{Country code mapping}
\usage{
country_to_iso3(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector matching the length of the input
}
\description{
Re-code country codes from iso2c to iso3c
}
\examples{
country_to_iso3(c("UK", "U.K.", "United Kingdom", "Great Britain & N.I.", "Greattt Britainnn"))
}
