% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{degrees_to_radians}
\alias{degrees_to_radians}
\title{Math functions}
\usage{
degrees_to_radians(x)
}
\arguments{
\item{x}{A numeric value or vector}
}
\value{
A numeric value or vector
}
\description{
Convert degrees to radians
}
\examples{
degrees_to_radians(45)
}
