% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envlogger.R
\name{env_header_val}
\alias{env_header_val}
\title{Get the value of a given field in an Envolgger file header}
\usage{
env_header_val(header, field_pattern, force_numeric = FALSE)
}
\arguments{
\item{header}{A tibble with columns \code{field} and \code{val} and all values in lower case}

\item{field_pattern}{A string that must match exactly one entry in the \code{field}column}

\item{force_numeric}{Logical (defaults to \code{FALSE}). If \code{TRUE}, the value is stripped of non-numeric characters and converted to numeric}
}
\value{
The value for the target field, as string or numeric (if \code{force_numeric = TRUE}).
}
\description{
Get the value of a given field in an Envolgger file header
}
\examples{
header <- tibble::tibble(field = c("field1", "field2", "field3"), val = c("a", "1", "a1"))
env_header_val(header, "field1")
env_header_val(header, "field2", force_numeric = TRUE)
env_header_val(header, "field2", force_numeric = TRUE)
}
