% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{explore_numeric_summary}
\alias{explore_numeric_summary}
\title{Title}
\usage{
explore_numeric_summary(dataset)
}
\arguments{
\item{dataset}{Explanation of 'dataset'}
}
\description{
Summary
}
\examples{

}
