% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.R
\name{file.ext}
\alias{file.ext}
\title{Get file extension from filename}
\usage{
file.ext(filename)
}
\arguments{
\item{filename}{character, Full file name}
}
\value{
File extension from filename
}
\description{
Get file extension from filename
}
\examples{
file.ext('a.b')

}

