% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{genLowerTriMatrix}
\alias{genLowerTriMatrix}
\title{A lower triangular matrix (incl. the diagonal) from a given vector.}
\usage{
genLowerTriMatrix(x, n)
}
\arguments{
\item{x}{the vector}

\item{n}{the number of rows/columns}
}
\value{
a lower triangular matrix with values corresponding to \code{x}
}
\description{
A lower triangular matrix (incl. the diagonal) from a given vector.
}
\examples{
 genLowerTriMatrix(1:10, 4)
}

