% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{genUpperTriMatrix}
\alias{genUpperTriMatrix}
\title{An upper triangular matrix (incl. the diagonal) from a given vector.}
\usage{
genUpperTriMatrix(x, n)
}
\arguments{
\item{x}{the vector}

\item{n}{the number of rows/columns}
}
\value{
an upper triangular matrix with values corresponding to \code{x}
}
\description{
An upper triangular matrix (incl. the diagonal) from a given vector.
}
\examples{
 genUpperTriMatrix(1:10, 4)
}

