% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hex.R
\name{get_hex}
\alias{get_hex}
\title{Extract HEX code by RGB}
\usage{
get_hex(x, y)
}
\arguments{
\item{x}{RGB codes in a \code{character} vector of length 1 (per color)}
}
\description{
This function extracts HEX codes from RGB codes
}
\examples{
x <- c("14 17 22")
hex <- sapply(strsplit(x, " "), function(x) rgb(x[1], x[2], x[3], maxColorValue=255))
}
\author{
Jeppe Vierø
}
