% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envlogger.R
\name{is.envlogger}
\alias{is.envlogger}
\title{Check if a path points to an EnvLogger file, logfile or none of the two}
\usage{
is.envlogger(path)
}
\arguments{
\item{path}{A path to a file}
}
\value{
A single numeric value, \code{1} if an EnvLogger file, \code{2} if an EnvLogger logfile, and \code{0} if none of the two.
}
\description{
Check if a path points to an EnvLogger file, logfile or none of the two
}
\examples{
path <- rtools_example()[1]
is.envlogger(path) # 1

path <- rtools_example("log_")[1]
is.envlogger(path)   # 2
}
