% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envlogger.R
\name{is.report_quick}
\alias{is.report_quick}
\title{Check if the paths provided point to EnvLogger files (quick)}
\usage{
is.report_quick(paths, negate = FALSE)
}
\arguments{
\item{paths}{A vector of paths to files.}

\item{negate}{If \code{TRUE}, inverts the resulting boolean vector.}
}
\value{
A logical vector the same length as \code{paths}.
}
\description{
Quickly determine if the paths provided point to EnvLogger report files.
This is done by testing if line 2 is equal to the string "www.electricblue.eu, Portugal".
As a result, reports generated using the the earliest versions of EnvLogger_Viewer will not be recognized.
Also, if the structure of the file has been changed (e.g., by opeining and saving in Excel), it will no longer be recognized as a report.
}
\examples{
paths <- dir(
 "/Users/ruiseabra/Dropbox/RS/bio/datasets/temp_loggers/reports/tidy/cctbon_norterocks/2021-01-18",
 full.names = TRUE,
 pattern = ".csv")
paths <- stringr::str_subset(paths, "log_", negate = TRUE)
is.envlogger(path) # 1
}
