% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_codes.R
\name{iso2_to_iso3}
\alias{iso2_to_iso3}
\title{Country code remapping}
\usage{
iso2_to_iso3(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector matching the length of the input
}
\description{
Re-code country codes from iso2c to iso3c
}
\examples{
iso2_to_iso3(c("FR", "DE", "RU"))
}
