% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_codes.R
\name{iso3_to_country}
\alias{iso3_to_country}
\title{Country code mapping}
\usage{
iso3_to_country(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector matching the length of the input
}
\description{
iso3c country code to country name
}
\examples{
iso3_to_country(c("GBR", "USA", "FRA"))
}
