% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_codes.R
\name{iso3_to_iso2}
\alias{iso3_to_iso2}
\title{Country code remapping}
\usage{
iso3_to_iso2(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector matching the length of the input
}
\description{
Re-code country codes from iso3c to iso2c
}
\examples{
iso3_to_iso2(c("FRA", "DEU", "RUS"))
}
