% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{klDiv}
\alias{klDiv}
\title{Kullback-Leibler divergence.}
\usage{
klDiv(p, q, X)
}
\arguments{
\item{p}{a density function of the form \code{p(X)} whose single argument is
a vector or matrix}

\item{q}{a density function of the same form as \code{p}}

\item{X}{a vector or matrix representing the observations}
}
\value{
the KL divergence between \code{p} and \code{q}
}
\description{
The Kullback-Leibler divergence between two distributions given their
density functions \code{p} and \code{q} along with samples \code{X}.
}
\details{
\deqn{
  KL(p || q) = \sum_{x \in \Omega} \log \frac{p(x)}{q(x)} p(x)
}
}

