% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{latlon_to_merc}
\alias{latlon_to_merc}
\title{Reproject points}
\usage{
latlon_to_merc(lon, lat)
}
\arguments{
\item{lon}{A numeric value or vector of longitude (East-West) values}

\item{lat}{A numeric value or vector of latitude (North-South) values}
}
\value{
A data.frame of Mercator coordinates
}
\description{
Convert Lat-Lons (WGS84) to Mercator
}
\examples{
latlon_to_merc(lon = 1, lat = 52)
}
