% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{legend.gradient}
\alias{legend.gradient}
\title{Legend for color gradient coded plots.}
\usage{
legend.gradient(tlx, tly, brx, bry, cols = heat.colors(5), labels = c(0, 1))
}
\arguments{
\item{tlx}{x-coordinate of the top-left corner}

\item{tly}{y-coordinate of the top-left corner}

\item{brx}{x-coordinate of the bottom-right corner}

\item{bry}{y-coordinate of the bottom-right corner}

\item{cols}{a vector or list of colors}

\item{labels}{a vector or list of labels}
}
\description{
Legend for color gradient coded plots.
}

