% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/lf.R
\name{lf}
\alias{lf}
\title{List the files in current local machine or on a remote host.}
\usage{
lf(path = ".", pattern = NULL, all.files = FALSE, full.names = FALSE,
  recursive = FALSE, ignore.case = FALSE, include.dirs = FALSE,
  no.. = FALSE)
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to the working directory, 'getwd()'.  Tilde expansion (see 'path.expand') is performed.  Missing values will be ignored. ":" is used to indicate a link to a remote host. The name of the host preceeds the semi-column.}

\item{pattern}{an optional regular expression.  Only file names which match the regular expression will be returned.}

\item{all.files}{a logical value.  If 'FALSE', only the names of visible files are returned.  If 'TRUE', all file names will be returned.}

\item{full.names}{a logical value.  If 'TRUE', the directory path is prepended to the file names to give a relative file path.  If 'FALSE', the file names (rather than paths) are returned.}

\item{recursive}{logical.  Should the listing recurse into directories? (not yet implemented for remote host)}

\item{ignore.case}{logical.  Should pattern-matching be case-insensitive? (not yet implemented for remote host)}

\item{include.dirs}{logical.  Should subdirectory names be included in recursive listings?  (They always are in non-recursive ones).}

\item{no..}{logical.  Should both '"."' and '".."' be excluded also from non-recursive listings? (automatic for remote host)}
}
\description{
This function is an extension of list.files that can also list the files on a remote computer.
}
\examples{
## On local machine, similarly to list.files
lf("~/user/data")
## On remote host,
lf("host:~/user/data")
}
\keyword{list,}
\keyword{remote}

