% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_database.R
\name{list_env_reports}
\alias{list_env_reports}
\title{List all EnvLogger reports in a folder (quick)}
\usage{
list_env_reports(folder)
}
\arguments{
\item{folder}{Path to a folder with Envlogger data}
}
\value{
A tibble with 3 columns containing the paths for all files inside \code{folder} that are (more or less) likely to be EnvLogger reports: \code{path} contains the full path for each file, \code{report} is a boolean vector containing the output from \code{is.report_quick} and \code{size} stores the size of each file (in bytes).
}
\description{
Given a path to a folder containing Envlogger data, list all files that likely are EnvLogger reports.
Since \code{is.report_quick} only tests if line 2 is equal to the string "www.electricblue.eu, Portugal", reports generated using the the earliest versions of EnvLogger_Viewer or where the structure of the file has been changed (e.g., by opeining and saving in Excel) are not directly identifiable. Therefore, all files that are (more or less) likely to be reports are returned, and the collumn \code{report} can be used for filtering based on the output from \code{is.report_quick}. The size of all files is also included, as it often is an important criterium for downstream processing.
}
\examples{
folder <- "/Users/ruiseabra/Dropbox/RS/bio/datasets/temp_loggers/reports"
list_env_reports(folder)
}
\seealso{
\code{\link[=is.report_quick]{is.report_quick()}}, \code{\link[=list_logs]{list_logs()}}
}
