% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_database.R
\name{list_logs}
\alias{list_logs}
\title{List all logfiles in a folder}
\usage{
list_logs(folder)
}
\arguments{
\item{folder}{Path to a folder with Envlogger data}
}
\value{
A vector with the paths to all logfiles found.
}
\description{
Given a path to a folder containing Envlogger data, list all the logfiles.
}
\examples{
folder <- "/Users/ruiseabra/Dropbox/RS/bio/datasets/temp_loggers/reports"
list_logs(folder)
}
\seealso{
\code{\link[=read_env_log]{read_env_log()}}, \code{\link[=read_deployment_info]{read_deployment_info()}}
}
