% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marchange.utils-package.R
\docType{package}
\name{marchange.utils-package}
\alias{marchange.utils}
\alias{marchange.utils-package}
\title{marchange.utils: Utility Tools for the MarChange Research Team (CIBIO-BIOPOLIS)}
\description{
Utility tools to get tide times and heights from festides2014, to read EnvLogger data and other. New tools added regularly.
}
\author{
\strong{Maintainer}: Rui Seabra \email{ptruimar@gmail.com} (\href{https://orcid.org/0000-0002-0240-3992}{ORCID})

Authors:
\itemize{
  \item Fernando Lima \email{fplima@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})
}

}
\keyword{internal}
