% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envlogger.R
\name{marchange_example}
\alias{marchange_example}
\title{Get paths to example files}
\usage{
marchange_example(pattern = NULL)
}
\arguments{
\item{pattern}{Pattern to select one or more example files. Pattern is vectorized, so more than one value can be supplied. If NULL, all example files are listed.}
}
\value{
The full path to one or more example files, or the filenames of all example files available.
}
\description{
\code{marchange.utils-package} comes bundled with several sample files in its inst/extdata directory. This function make them easy to access
}
\examples{
# Get the filenames of all example files
marchange_example()

# Get the full path to a target example file
marchange_example("04FD 1700 8A6C 05-20230830 180125")

# Get the full paths to any example files matching a search string
marchange_example("2023-08-30 - uksca")

# 'marchange_example()' is vectorized, meaning that multiple search strings can be used
marchange_example(c("04E9 6B00 943C 0C", "04CA A900 E31C 0F"))
}
