% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{merc_to_latlon}
\alias{merc_to_latlon}
\title{Reproject points}
\usage{
merc_to_latlon(x, y)
}
\arguments{
\item{lon}{A numeric value or vector of Mercator East-West values}

\item{lat}{A numeric value or vector of Mercator North-South values}
}
\value{
A data.frame of WGS84 coordinates
}
\description{
Convert Mercator to Lat-Lons (WGS84)
}
\examples{
merc_to_latlon(lon = 1, lat = 52)
}
