% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{merc_to_wmt}
\alias{merc_to_wmt}
\title{Web Map Tiles}
\usage{
merc_to_wmt(x, y, z)
}
\arguments{
\item{x}{A numeric value or vector}
}
\value{
A numeric value or vector
}
\description{
Convert Mercator to Web Map Tile coordinates
}
\examples{
m_ext = pi * 6378137 # Mercator extent
dat = dplyr::tibble(x = pi * 6378137 * runif(50, min=-1), y = pi * 6378137 * runif(50, min=-1))
merc_to_wmt(dat$x, dat$y, 4)
}
