% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{mutualInformation}
\alias{mutualInformation}
\title{Mutual information of a random variable and corresponding class attribute.}
\usage{
mutualInformation(x, y, breaks, ...)
}
\arguments{
\item{x}{a vector or a factor}

\item{y}{a factor of the same length as x that encodes the class of each
corresponding sample of x}

\item{breaks}{see \code{\link{cut}}}

\item{...}{arguments passed on to \code{\link{entropy}}}
}
\value{
the mutual information of X and Y
}
\description{
\deqn{I(X;Y) = H(X) - H(X|Y)}
}
\seealso{
\code{\link{entropy}} and \code{\link{entropyGivenClass}}
}

