% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.R
\name{nca.grad.kl}
\alias{nca.grad.kl}
\title{The gradient of A corresponding to KL divergence.}
\usage{
nca.grad.kl(X, A, P, klass)
}
\arguments{
\item{X}{a matrix whose rows represent the original data}

\item{A}{the current transformation matrix}

\item{P}{a matrix where \eqn{P_ij} denotes the probability of point
\eqn{X_i} belonging to the same class a \eqn{X_j}}

\item{klass}{a vector or factor whose i-th element corresponds to the class
of point \eqn{X_i}}
}
\value{
the gradient of A corresponding to KL divergence
}
\description{
Maximize the probability of obtaining an error-free classification of the
entire training set.
}
\references{
Goldberger et al., "Neighbourhood components analysis", 2004
}

