% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.R
\name{nca}
\alias{nca}
\title{Neighbourhood Component Analysis}
\usage{
nca(X, klass, d = 2, maxIter = 100, lambda = 0.001, epsilon = 0.01,
  algorithm = "points", verb = T)
}
\arguments{
\item{X}{a matrix whose rows represent the original data}

\item{klass}{a vector or factor whose i-th element corresponds to the class
of point \eqn{X_i}}

\item{d}{the dimensionality for the result}

\item{maxIter}{the maximum number of iterations}

\item{lambda}{the learning rate (usually <<1)}

\item{epsilon}{the desired accuracy, i.e. once the Frobenius norm of the
gradient becomes less than or equal to \code{epsilon}, convergence is assumed}

\item{algorithm}{one of \code{c("points", "kl")}}

\item{verb}{\verb{TRUE} for verbose output, else \verb{FALSE}}
}
\value{
a linear transformation matrix that maps the original data in \code{X}
 to a \code{d}-dimensional space
}
\description{
Neighbourhood Component Analysis
}

