\name{p.disease}

\alias{p.disease}

\title{Caclulate disease frequency}

\description{Calculate the disease allele frequency given values for
  frequency and OR}

\usage{p.disease(p, OR)}

\arguments{
  \item{p}{Allele frequency in controls.}
  \item{OR}{Odds Ratio.}
}

\value{Allele frequency in cases.}

\author{Randall Johnson}

\keyword{htest}

