% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{parse_hash}
\alias{parse_hash}
\title{Shiny functions}
\usage{
parse_hash(hash_str)
}
\arguments{
\item{hash_str}{A URL hash string passed to Shiny}
}
\value{
A list of key/value pairs
}
\description{
Parse hash: parser to turn Shiny's session$clientData$url_hash input into a list of key/value pairs
}
\examples{
observeEvent(session$clientData$url_hash, {
    hash = parse_hash(session$clientData$url_hash)
    ...
}
}
