% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{partial}
\alias{partial}
\title{A partially applied function object.}
\usage{
partial(FUN, ...)
}
\arguments{
\item{FUN}{a function object}

\item{...}{the parameters to be bound to the function object}
}
\value{
a function object with a \code{...} parameter list where all of the
 already given parameters \code{...} have already been bound
}
\description{
A partially applied function object.
}
\examples{
 sumOfTwoAndWhatever <- partial(sum, 2)
 sumOfTwoAndWhatever(4) == 6
 sumOfTwoAndWhatever(4, 7) == 13
}

