% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition.array}
\alias{partition.array}
\title{Partition the given array \code{X} into \code{numPartitions} partitions.}
\usage{
\method{partition}{array}(X, numPartitions, axis = 1, ...)
}
\arguments{
\item{X}{an instance of a supported class}

\item{numPartitions}{the number of partitions}

\item{axis}{the axis along which to partition}

\item{...}{reserved}
}
\value{
a list containing the partitions of \code{X}
}
\description{
Partition the given array \code{X} into \code{numPartitions} partitions.
}

