% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition.numeric}
\alias{partition.numeric}
\title{Partition the given vector \code{X} into \code{numPartitions} partitions.}
\usage{
\method{partition}{numeric}(X, numPartitions, ...)
}
\arguments{
\item{X}{an instance of a supported class}

\item{numPartitions}{the number of partitions}

\item{...}{reserved}
}
\value{
a list containing the partitions of \code{X}
}
\description{
Partition the given vector \code{X} into \code{numPartitions} partitions.
}

