% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{S3 method for partitioning \code{X} into \code{numPartitions} partitions.}
\usage{
partition(X, numPartitions, ...)
}
\arguments{
\item{X}{an instance of a supported class}

\item{numPartitions}{the number of partitions}

\item{...}{passed on to S3 method delegates}
}
\value{
a list containing the partitions of \code{X}
}
\description{
S3 method for partitioning \code{X} into \code{numPartitions} partitions.
}

