% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.full}
\alias{pca.full}
\title{Principal Component Analysis of the given dataset \code{X}.}
\usage{
pca.full(X, center = TRUE, retX = TRUE, tol = 1)
}
\arguments{
\item{X}{a matrix where each row denotes one multivariate sample}

\item{center}{a boolean indicating whether the data should be centered}

\item{retX}{whether to compute and return a projection of the (possibly
centered) original dataset onto its principal components}

\item{tol}{a value in [0,1] indicating which components to keep}
}
\value{
a list consisting of
  \item{\code{X}}{if \code{retX} is \code{TRUE}, a projection of the (possibly
    centered original dataset onto its principal components, else \code{NULL}}
  \item{\code{V}}{an orthonormal matrix whose columns resemble the
    principal components, i.e. the eigenvectors of the covariance matrix}
  \item{\code{sd}}{the standard deviations of the principal components, i.e.
    the square roots of the eigenvalues of the covariance matrix}
  \item{\code{mean}}{if \code{center} is \code{TRUE}, the mean of the original
    dataset, else \code{NULL}}
}
\description{
PCA is performed through Singular Value Decomposition as that is numerically
more stable than doing it by means of eigenvalue decomposition of the
covariance matrix.
}
\details{
Depending on the given tolerance \code{tol}, either all principal components
are computed, i.e. the trace of the transformed covariance matrix is the same
as before the transformation (no information is lost), or only those
components whose variances sum up to at least a fraction of \code{tol}
of the sum of the total variance.
}

