% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot_prob}
\alias{plot_prob}
\title{Function to plot probabilities}
\usage{
plot_prob(x, y, ints = 4, binom.model = NULL, ylim = NULL, xlab = "",
  ylab = "")
}
\arguments{
\item{x}{x-values}

\item{y}{y-values}

\item{ints}{Number of intervals (defauls is 4)}

\item{binom.model}{Not yet implemented. Will be used to plot the predicted probability from binomila model}

\item{ylim}{Specify ylim for plot}
}
\description{
Summarizes probabilities per interval, with the size of each circle proportional to sample size
}
\examples{

}
