% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotAllVsAll}
\alias{plotAllVsAll}
\title{2d histogram plots of all variables vs. each other.}
\usage{
plotAllVsAll(dataset, nbins = 200, oma = c(2, 1, 3.5, 1), col = c("white",
  heat.colors(16)))
}
\arguments{
\item{dataset}{a dataframe or matrix whose rows contain the samples and
whose columns denote the variables}

\item{nbins}{an integer or a two-vector describing the number of bins per axis,
passed on to \code{hist2d}}

\item{oma}{passed on to \code{par}}

\item{col}{colors of the various levels, passed on to \code{hist2d}}
}
\description{
2d histogram plots of all variables vs. each other.
}

