% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotStdDev}
\alias{plotStdDev}
\title{Plot standard deviation markers.}
\usage{
plotStdDev(x, y, stddev, length = 0.1, ...)
}
\arguments{
\item{x}{x-coordinates of points}

\item{y}{y-coordinates of points}

\item{stddev}{standard deviations corresponding to \code{x} and \code{y}}

\item{length}{the length of the perpendicular bar, i.e. top and bottom of thex
marker}

\item{...}{passed on to \code{arrows}}
}
\description{
Plot standard deviation markers.
}

