% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{powerSet}
\alias{powerSet}
\title{The power set of the given \code{x}, minus the empty set.}
\usage{
powerSet(x)
}
\arguments{
\item{x}{a set of values}
}
\value{
the power set of the given \code{x}, minus the empty set.
}
\description{
The power set of the given \code{x}, minus the empty set.
}

