% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{print.confusionMatrix}
\alias{print.confusionMatrix}
\title{Print the given confusion matrix.}
\usage{
\method{print}{confusionMatrix}(x, ...)
}
\arguments{
\item{x}{an instance of \code{confusionMatrix}}

\item{...}{ignored}
}
\description{
Print the given confusion matrix along with information on the overall
accuracy as well as precision and recall for each class.
}

