% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{private__as_year_qtr_format}
\alias{private__as_year_qtr_format}
\title{only returns a value for dates matching quarter so that when used with scale_x_date there are no axis
ticks for anything other than quarters}
\usage{
private__as_year_qtr_format(x)
}
\description{
only returns a value for dates matching quarter so that when used with scale_x_date there are no axis
ticks for anything other than quarters
}
