% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwSquaredDist.R
\name{pwSquaredDist}
\alias{pwSquaredDist}
\title{Squared distance of each pair of row vectors of the matrix \code{X}.}
\usage{
pwSquaredDist(X, asMatrix = FALSE)
}
\arguments{
\item{X}{a matrix}

\item{asMatrix}{a boolean indicating whether the result should be a
vector or a symmetric matrix, defaults to \code{FALSE}}
}
\value{
a vector of length N*(N-1)/2 if \code{asMatrix=T}, else a NxN
  symmetric matrix
}
\description{
Squared distance of each pair of row vectors of the matrix \code{X}.
}
\examples{
Y <- matrix(c(1, -2, 3, 55, -8, 4, 3, 7, 11), ncol=3)
distances <- pwSquaredDist(Y)        # Result as vector

D <- pwSquaredDist(Y, asMatrix=TRUE) # Result as lower triangular matrix
}

