% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{radians_to_degrees}
\alias{radians_to_degrees}
\title{Math functions}
\usage{
radians_to_degrees(x)
}
\arguments{
\item{x}{A numeric value or vector}
}
\value{
A numeric value or vector
}
\description{
Convert radians to degrees
}
\examples{
radians_to_degrees(1)
}
