% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_database.R
\name{read_deployment_info}
\alias{read_deployment_info}
\title{Read Envlogger logfiles and list all serials}
\usage{
read_deployment_info(paths, col_names = FALSE)
}
\arguments{
\item{paths}{A vector of paths to Envlogger logfiles}

\item{col_names}{Logical, whether to include the names of columns in the \code{.csv} output file.}
}
\value{
A tibble with one row for each unique serial name and 5 columns with relevant metadata.
}
\description{
Given a vector of paths to Envlogger logfiles, read and list the serials included.
}
\examples{
folder <- "/Users/ruiseabra/Dropbox/RS/bio/datasets/temp_loggers/reports"
paths  <- rtools_example("/log")
read_deployment_info(paths)
}
\seealso{
\code{\link[=read_env_log]{read_env_log()}}
}
