% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envlogger.R
\name{read_env}
\alias{read_env}
\title{Read an EnvLogger file or logfile}
\usage{
read_env(path, zero_secs = TRUE)
}
\arguments{
\item{path}{Path to an EnvLogger file or logfile}

\item{zero_secs}{Logical, whether to remove trailing seconds (only used if path points to an EnvLogger file; adjustment determined based on the first timestamp)}
}
\value{
If path points to an EnvLogger file, a tibble with 1 row and 13 columns, among which \code{id} and \code{data}. If path points to an EnvLogger logfile, a tibble with one row for each interaction with an EnvLogger and 15 columns with relevant metadata.
}
\description{
Given a file path, check if it points to an EnvLogger file or logfile and, if \code{TRUE}, read it.
}
\examples{
path <- rtools_example()[3]
read_env(path, zero_secs = TRUE) # an EnvLogger file

path <- rtools_example("log_")[1]
read_env(path) # an EnvLogger logfile
}
\seealso{
\code{\link[=read_env_header]{read_env_header()}}, \code{\link[=read_env_data]{read_env_data()}}, \code{\link[=read_env_log]{read_env_log()}}, \code{\link[=plot_env]{plot_env()}}
}
