% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_database.R
\name{read_env_details}
\alias{read_env_details}
\title{List all EnvLogger reports in a folder (quick) and grab their details}
\usage{
read_env_details(
  folder,
  min_size = 1000,
  max_size = 9999999999,
  parallel = FALSE
)
}
\arguments{
\item{folder}{Path to a folder with Envlogger data}

\item{min_size}{A numeric value (in bytes); only files larger than this will be retained; defaults to 1000 so that reports without any temperature data (or with only a few readings) are excluded.}

\item{max_size}{A numeric value (in bytes); only files smaller than this will be retained}

\item{parallel}{Logical, whether or not to use parallel computing (defaults to FALSE; based on the \code{future.apply} package)}
}
\value{
A tibble with 3 columns containing the paths for all files inside \code{folder} that are (more or less) likely to be EnvLogger reports: \code{path} contains the full path for each file, \code{report} is a boolean vector containing the output from \code{is.report_quick} and \code{size} stores the size of each file (in bytes).
}
\description{
Given a path to a folder containing Envlogger data, list all files that likely are EnvLogger reports and grab their header details.
Read the description of \code{\link[=is.report_quick]{is.report_quick()}} to better understand the limitations of the way files are quickly assessed as reports or not.
}
\examples{
folder <- "/Users/ruiseabra/Dropbox/RS/bio/datasets/temp_loggers/reports"
read_env_details(folder, parallel = TRUE)
}
\seealso{
\code{\link[=is.report_quick]{is.report_quick()}}, \code{\link[=list_env_reports]{list_env_reports()}}
}
