% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envlogger.R
\name{read_env_header}
\alias{read_env_header}
\title{Read the header of an EnvLogger file}
\usage{
read_env_header(path, check = TRUE)
}
\arguments{
\item{path}{Path to an EnvLogger file}

\item{check}{Logical, whether or not to check if path points to an EnvLogger file. Defaults to \code{TRUE}. Only meant to be \code{FALSE} when called from within \code{read_env()}, so as not to repeat the same check multiple times. \strong{Do not set it to \code{FALSE} when running this function directly.}
#'}
}
\value{
A tibble with 1 row and 12 columns with the metadata values for most of the fields recorded in the header.
}
\description{
Given a file path, check if it points to an EnvLogger file and, if \code{TRUE}, read just the header
}
\examples{
path <- rtools_example()[3]
read_env_header(path)

paths <- rtools_example("iemin_iespi")
paths <- paths[!grepl("log_", paths)]
headers <- purrr::map(paths, read_env_header)
purrr::list_rbind(headers)
}
\seealso{
\code{\link[=read_env_data]{read_env_data()}}, \code{\link[=read_env]{read_env()}}, \code{\link[=plot_env]{plot_env()}}
}
