% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{report_dependencies}
\alias{report_dependencies}
\title{Interactive tools}
\usage{
report_dependencies(path = "./", print = FALSE, simple = TRUE)
}
\arguments{
\item{path}{Character string of directory to report}

\item{print}{Logical whether to print output to console}

\item{simple}{Logical whether to return simple vector or pre-loaded/invoked breakdown}
}
\value{
A vector (simple=TRUE) or list (simple=FALSE)
}
\description{
Report all explicit R dependencies in current directory (recursive)
}
\examples{
report_dependencies()
report_dependencies('~/Downloads/')
report_dependencies() |> tools::package_dependencies(reverse = FALSE)
}
