% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{rt__mock__attribution_to_clickstream}
\alias{rt__mock__attribution_to_clickstream}
\title{transforms .campaign_data into structure that is based on events, meaning the conversion would be
a specific event (e.g. clicking a submit button on a website)
this mimics what we would typically see from a clickstream
randomly assigns "Button Submit Event" & "Button Submit Event 2"
randomly gives the conversion timestamp either the same timestamp of the corresponding step or 1 second after}
\usage{
rt__mock__attribution_to_clickstream(.campaign_data)
}
\arguments{
\item{.campaign_data}{dataframe with id|timestamp|step|step_type|num_conversions|conversino_value columns
This dataframe has a value num_conversions on the step association with the conversion.
So for example, if a user signs up on the pricing page, there is no specific signup step/event,
but rather a step for the pricing page indicating number of conversions (probably always 1 for a signup event)
and total value of conversions}
}
\description{
transforms .campaign_data into structure that is based on events, meaning the conversion would be
a specific event (e.g. clicking a submit button on a website)
this mimics what we would typically see from a clickstream
randomly assigns "Button Submit Event" & "Button Submit Event 2"
randomly gives the conversion timestamp either the same timestamp of the corresponding step or 1 second after
}
