% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_add_year_month_factors}
\alias{rt_add_year_month_factors}
\title{as `[.column]_year` and `[.column]_month` factor variables to .data based off of date column .column}
\usage{
rt_add_year_month_factors(.data, .column, .abbreviate = FALSE)
}
\arguments{
\item{.data}{date vector}

\item{.column}{name of date column in .data}

\item{.abbreviate}{use month abbreviations}
}
\description{
as `[.column]_year` and `[.column]_month` factor variables to .data based off of date column .column
}
