% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{rt_are_dataframes_equal}
\alias{rt_are_dataframes_equal}
\title{compares two dataframes and returns TRUE if they are both equal}
\usage{
rt_are_dataframes_equal(dataframe1, dataframe2)
}
\arguments{
\item{dataframe1}{a dataframe to compare}

\item{dataframe2}{a dataframe to compare}
}
\description{
compares two dataframes and returns TRUE if they are both equal
}
