% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rt_are_dataframes_equal_from_file}
\alias{rt_are_dataframes_equal_from_file}
\title{compares two dataframes, one of them will be loaded from the `file`}
\usage{
rt_are_dataframes_equal_from_file(dataframe1, rds_file)
}
\arguments{
\item{dataframe1}{a dataframe to compare}

\item{rds_file}{a file c}
}
\description{
compares two dataframes, one of them will be loaded from the `file`
}
