% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_are_numerics_equal}
\alias{rt_are_numerics_equal}
\title{Checks if two numbers are equivalent out to `n` decimals where `n` is the `num_decimals` parameter}
\usage{
rt_are_numerics_equal(n1, n2, num_decimals)
}
\arguments{
\item{n1}{number or vector of numbers}

\item{n2}{number or vector of numbers}

\item{num_decimals}{round each value `n` decimals where `n` is `num_decimals`}
}
\description{
Checks if two numbers are equivalent out to `n` decimals where `n` is the `num_decimals` parameter
}
