% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{rt_attribution_pivot_longer}
\alias{rt_attribution_pivot_longer}
\title{transforms attribution dataframe to long format}
\usage{
rt_attribution_pivot_longer(attribution_models)
}
\arguments{
\item{attribution_models}{dataframe with columns `channel_name | xxx_conversions | xxx_value`}
}
\description{
transforms attribution dataframe to long format
}
