% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{rt_campaign_add_path_id}
\alias{rt_campaign_add_path_id}
\title{adds `.path_id` column to dataframe. Each conversion step triggers a new path-id for subsequent steps}
\usage{
rt_campaign_add_path_id(
  .campaign_data,
  .id = "id",
  .timestamp = "timestamp",
  .num_conversions = "num_conversions",
  .conversion_value = "conversion_value",
  .use_first_conversion = TRUE,
  .sort = TRUE
)
}
\arguments{
\item{.campaign_data}{dataframe with columns: `id | timestamp | step | num_conversions | conversion_value`}

\item{.id}{string identifying the id column}

\item{.timestamp}{string identifying the timestamp column}

\item{.num_conversions}{string identifying the num_conversions column}

\item{.conversion_value}{string identifying the conversion_values column}

\item{.use_first_conversion}{if true, only use the first conversion - the path id will be identical to the .id (by definition there will never be more than 1 path)}

\item{.sort}{if true, the df is sorted by id, timestamp, conversion_values, & num_conversions; the dataframe needs to be sorted in this way to work, but this option let's the user avoid this action if it has already been done}
}
\description{
adds `.path_id` column to dataframe. Each conversion step triggers a new path-id for subsequent steps
}
