% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{rt_clickstream_to_attribution}
\alias{rt_clickstream_to_attribution}
\title{transforms .clickstream_data into the expected format for attribution calculations
different types of conversion events will be ignored, so the user is expected to filter out any
conversion events they are not interested in.}
\usage{
rt_clickstream_to_attribution(.clickstream_data)
}
\arguments{
\item{.clickstream_data}{dataframe with id|timestamp|step|step_type|num_conversions|conversion_value columns
This dataframe has "clickstream" data, which means that it has a list of steps/events that might
     correspond to, for example, page visits on a website.
num_conversions should indicate which steps are conversion events.
A conversion event should be its own step, that has a timestamp equal to or after the step that
should get the conversion event.
    For example, if someone visits the pricing page, and then signs up for the product (which is the conversion),
    there should be a single row (i.e. step) for the visit to the pricing page, and a single row for the conversion step.
    The timestamp of the conversion event, in this case, would be immdediately after (seconds or minutes) the pricing step.
    The step that is before the conversion event (regardless of how much before) gets credit (last-touch) for the conversion.}
}
\description{
This function does not handle the case when the there are multiple types of conversions that are
triggered from a single step
For example: Someone lands on the homepage and from that page does 2 conversions `Lead Form-Fill` & `Lead Sign-up`)
The user is expected to filter the events of interest accordingly.
}
