% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_colors}
\alias{rt_colors}
\title{Hex colors vector}
\usage{
rt_colors(color_names = NULL, sets = NULL, return_named_vector = FALSE)
}
\arguments{
\item{color_names}{filter by the names of the colors}

\item{sets}{filter by the set index}

\item{return_named_vector}{returns the colors as a named vector}
}
\description{
Hex colors vector
}
