% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_colors_good_bad}
\alias{rt_colors_good_bad}
\title{Hex colors vector representing "good" (green) and "bad" (red)}
\usage{
rt_colors_good_bad(good_first = TRUE)
}
\arguments{
\item{good_first}{if `TRUE` then index 1 is the hex color corresponding to `good` and
index 2 corresponds to bad, otherwise the reverse.}
}
\description{
Hex colors vector representing "good" (green) and "bad" (red)
}
