% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_difftime_numeric}
\alias{rt_difftime_numeric}
\title{returns numeric value of difftime (date_last - date_first)}
\usage{
rt_difftime_numeric(date_last, date_first, units = "days")
}
\arguments{
\item{date_last}{the date to subtract from}

\item{date_first}{the date to subtract}

\item{units}{same values of difftime, defaults to `days`}
}
\description{
returns numeric value of difftime (date_last - date_first)
}
