% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_equal_include_na}
\alias{rt_equal_include_na}
\title{There are times when you want to filter/match any value, *including* NA
So if}
\usage{
rt_equal_include_na(x, y)
}
\arguments{
\item{x}{single value or vector to compare with y}

\item{y}{single value or vector to compare with x}
}
\description{
There are times when you want to filter/match any value, *including* NA
So if
}
