% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_plot_aggregate_2_numerics}
\alias{rt_explore_plot_aggregate_2_numerics}
\title{returns a eih}
\usage{
rt_explore_plot_aggregate_2_numerics(dataset, variable,
  comparison_variable, aggregation_function = NULL,
  aggregation_function_name = NULL, aggregation_count_minimum = 30,
  show_resampled_confidence_interval = FALSE, show_points = FALSE,
  show_labels = FALSE, x_zoom_min = NULL, x_zoom_max = NULL,
  y_zoom_min = NULL, y_zoom_max = NULL, base_size = 11)
}
\arguments{
\item{dataset}{dataframe}

\item{variable}{a numeric variable (y-axis)}

\item{comparison_variable}{the additional numeric variable (x-axis) that will be grouped}

\item{aggregation_function}{the aggregation function that gives, for example,
the mean of the `variable` grouped by the `comparison_variable`
if no aggregation function is supplied the boxplots are showed for each grouped `comparison_variable`
value.}

\item{aggregation_function_name}{the friendly name of the `aggregation_function`}

\item{aggregation_count_minimum}{the minimum number of samples to included for each `comparison_variable`
value. The default is 30, so any value of `comparison_variable` that didn't occur at least 30 times
would be removed. We need at least 30, otherwise when we bootstrap resample e.g. with a group that has
1 sample we'd pull e.g. 100 random samples of the same value}

\item{show_resampled_confidence_interval}{when `aggregation_function` is not null, then we have the option
to show a confidence interval based on resampling}

\item{show_points}{the show points}

\item{show_labels}{show the labels}

\item{x_zoom_min}{adjust (i.e. zoom in) to the x-axis; sets the minimum x-value for the adjustment}

\item{x_zoom_max}{adjust (i.e. zoom in) to the x-axis; sets the maximum x-value for the adjustment}

\item{y_zoom_min}{adjust (i.e. zoom in) to the y-axis; sets the minimum y-value for the adjustment}

\item{y_zoom_max}{adjust (i.e. zoom in) to the y-axis; sets the maximum y-value for the adjustment}

\item{base_size}{uses ggplot's base_size parameter for controling the size of the text}
}
\description{
returns a eih
}
