% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_plot_boxplot}
\alias{rt_explore_plot_boxplot}
\title{returns a boxplot of `variable`, potentally grouped by `comparison_variable`}
\usage{
rt_explore_plot_boxplot(dataset, variable, comparison_variable = NULL,
  color_variable = NULL, facet_variable = NULL, y_zoom_min = NULL,
  y_zoom_max = NULL, base_size = 11)
}
\arguments{
\item{dataset}{dataframe containing numberic columns}

\item{variable}{the variable from which to create a boxplot}

\item{comparison_variable}{the additional variable to group by; must be a string/factor column}

\item{color_variable}{if comparison_variable is used, then if color_variable is used the boxplots will be colored by this variable rather than each of the comparison_variable categories. `color_variable` must be categoric}

\item{facet_variable}{if comparison_variable is used, then if facet_variable is used the boxplots will be faceted by this variable `faceted` must be categoric}

\item{y_zoom_min}{adjust (i.e. zoom in) to the y-axis; sets the minimum y-value for the adjustment}

\item{y_zoom_max}{adjust (i.e. zoom in) to the y-axis; sets the maximum y-value for the adjustment}

\item{base_size}{uses ggplot's base_size parameter for controling the size of the text}
}
\description{
returns a boxplot of `variable`, potentally grouped by `comparison_variable`
}
