% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_plot_categoric_heatmap}
\alias{rt_explore_plot_categoric_heatmap}
\title{returns a heat map with histogram on top/side showing distribution of both features.}
\usage{
rt_explore_plot_categoric_heatmap(dataset, x_variable, y_variable,
  sum_by_variable = NULL, count_distinct_variable = NULL,
  multi_value_delimiter = NULL, base_size = 11)
}
\arguments{
\item{dataset}{dataframe containing numberic columns}

\item{x_variable}{the variable (e.g. factor) to get unique values from}

\item{y_variable}{the additional variable to group by; must be a string/factor column}

\item{sum_by_variable}{the numeric variable to sum}

\item{count_distinct_variable}{when aggregating, rather than counting the total number of records, count distinct occurances of this variabled (cannot be used with `sum_by_variable`)}

\item{multi_value_delimiter}{if the variable contains multiple values (e.g. "A", "A, B", ...) then setting
this variable to the delimiter will cause the function to count seperate values}

\item{base_size}{uses ggplot's base_size parameter for controling the size of the text}
}
\description{
If `multi_value_delimiter` is not NULL, then it counts all the values found after it splits/separates the
     variable by the delimiter. If `sum_by_variable` is NULL, it counts the values and the denominator for
     the `percent` column returned is the total number of records. If `sum_by_variable` is not NULL, then
     when multiple values are found, each value is weighted by the value found in `sum_by_variable`, and
     the denominator for the `percent` column returned is the `sum` of `sum_by_variable` (before the values
     are split).
}
\details{
Currently only works when using only `variable` (not `comparison_variable`)
}
