% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_plot_categoric_numeric_aggregation}
\alias{rt_explore_plot_categoric_numeric_aggregation}
\title{returns a graph for categoric/numeric variables based on the aggregation_type}
\usage{
rt_explore_plot_categoric_numeric_aggregation(
  dataset,
  categoric_variable,
  numeric_variable,
  aggregation_type,
  color_variable = NULL,
  facet_variable = NULL,
  show_variable_totals = TRUE,
  show_comparison_totals = TRUE,
  simple_mode = FALSE,
  base_size = 11
)
}
\arguments{
\item{dataset}{dataframe containing numberic columns}

\item{categoric_variable}{the variable (e.g. factor) to get unique values from}

\item{numeric_variable}{the additional variable to group by; must be a string/factor column}

\item{aggregation_type}{must be one of 'Total', 'Mean', 'Average Value Per Record', 'Median', 'Boxplot'}

\item{color_variable}{additional categoric variable to color by}

\item{facet_variable}{additional categoric variable to facet by}

\item{show_variable_totals}{if TRUE (the default) the graph will display the totals for the variable}

\item{show_comparison_totals}{if TRUE (the default) the graph will display the totals for the comparison_variable}

\item{simple_mode}{simplifies the coloring scheme and text}

\item{base_size}{uses ggplot's base_size parameter for controling the size of the text}
}
\description{
returns a graph for categoric/numeric variables based on the aggregation_type
}
