% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_plot_histogram}
\alias{rt_explore_plot_histogram}
\title{returns a histogram of `variable`}
\usage{
rt_explore_plot_histogram(dataset, variable, comparison_variable = NULL,
  num_bins = 30, x_zoom_min = NULL, x_zoom_max = NULL,
  base_size = 11)
}
\arguments{
\item{dataset}{dataframe containing numberic columns}

\item{variable}{the variable from which to create a histogram}

\item{comparison_variable}{(optional) the additional variable to group by; must be a string/factor column}

\item{num_bins}{the number of bins that the histogram will use}

\item{x_zoom_min}{adjust (i.e. zoom in) to the x-axis; sets the minimum x-value for the adjustment}

\item{x_zoom_max}{adjust (i.e. zoom in) to the x-axis; sets the maximum x-value for the adjustment}

\item{base_size}{uses ggplot's base_size parameter for controling the size of the text}
}
\description{
returns a histogram of `variable`
}
