% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_plot_numeric_heatmap}
\alias{rt_explore_plot_numeric_heatmap}
\title{returns a heat map with histogram on top/side showing distribution of both features.}
\usage{
rt_explore_plot_numeric_heatmap(dataset, x_variable, y_variable,
  n_cuts = 6, x_cut_sequence = NULL, y_cut_sequence = NULL,
  sum_by_variable = NULL, count_distinct_variable = NULL,
  multi_value_delimiter = NULL, base_size = 11)
}
\arguments{
\item{dataset}{dataframe containing numberic columns}

\item{x_variable}{the numeric variable to show on the x-axis}

\item{y_variable}{the numeric variable to show on the y-axis}

\item{n_cuts}{the number of cut points (will generate `n_cuts`-1 groups)}

\item{x_cut_sequence}{a vector of cut points; if not NULL, overrides `n_cuts` for the x variable}

\item{y_cut_sequence}{a vector of cut points; if not NULL, overrides `n_cuts` for the y variable}

\item{sum_by_variable}{the numeric variable to sum}

\item{count_distinct_variable}{when aggregating, rather than counting the total number of records, count distinct occurances of this variabled (cannot be used with `sum_by_variable`)}

\item{multi_value_delimiter}{(currently not implemented) if the variable contains multiple values (e.g. "A", "A, B", ...) then setting
this variable to the delimiter will cause the function to count seperate values}

\item{base_size}{uses ggplot's base_size parameter for controling the size of the text}
}
\description{
returns a heat map with histogram on top/side showing distribution of both features.
}
