% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_plot_scatter}
\alias{rt_explore_plot_scatter}
\title{returns a scatterplot of `variable` (numeric, x-axis) compared against `comparison_variable` (y-axis)}
\usage{
rt_explore_plot_scatter(dataset, variable, comparison_variable = NULL,
  alpha = 0.3, x_zoom_min = NULL, x_zoom_max = NULL,
  y_zoom_min = NULL, y_zoom_max = NULL, base_size = 11)
}
\arguments{
\item{dataset}{dataframe containing numberic columns}

\item{variable}{the variable from which to create a boxplot}

\item{comparison_variable}{the additional variable to group by; must be a string/factor column}

\item{alpha}{controls transparency}

\item{x_zoom_min}{adjust (i.e. zoom in) to the x-axis; sets the minimum x-value for the adjustment}

\item{x_zoom_max}{adjust (i.e. zoom in) to the x-axis; sets the maximum x-value for the adjustment}

\item{y_zoom_min}{adjust (i.e. zoom in) to the y-axis; sets the minimum y-value for the adjustment}

\item{y_zoom_max}{adjust (i.e. zoom in) to the y-axis; sets the maximum y-value for the adjustment}

\item{base_size}{uses ggplot's base_size parameter for controling the size of the text}
}
\description{
returns a scatterplot of `variable` (numeric, x-axis) compared against `comparison_variable` (y-axis)
}
