% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_plot_time_series}
\alias{rt_explore_plot_time_series}
\title{returns a time-series plot}
\usage{
rt_explore_plot_time_series(dataset, variable,
  comparison_variable = NULL, comparison_function = NULL,
  comparison_function_name = NULL, color_variable = NULL,
  y_zoom_min = NULL, y_zoom_max = NULL, show_points = FALSE,
  show_labels = FALSE, base_size = 11)
}
\arguments{
\item{dataset}{dataframe}

\item{variable}{a variable (x-axis) that is a date type}

\item{comparison_variable}{the additional numeric variable (y-axis)}

\item{comparison_function}{if a comparison variable is supplied, a function must be given so the plot knows how to graph it (e.g. sum, mean, median)}

\item{comparison_function_name}{name of the function so that it can be plotted on the Y-axis label}

\item{color_variable}{an optional variable (categoric) that seperates the time series}

\item{y_zoom_min}{adjust (i.e. zoom in) to the y-axis; sets the minimum y-value for the adjustment}

\item{y_zoom_max}{adjust (i.e. zoom in) to the y-axis; sets the maximum y-value for the adjustment}

\item{show_points}{if TRUE adds points to the graph}

\item{show_labels}{if TRUE adds labels to each point}

\item{base_size}{uses ggplot's base_size parameter for controling the size of the text}
}
\description{
returns a time-series plot
}
