% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_plot_unique_values}
\alias{rt_explore_plot_unique_values}
\title{returns a barchart of the unique value counts for a given dataset/variable}
\usage{
rt_explore_plot_unique_values(dataset, variable, order_by_count = TRUE,
  base_size = 11)
}
\arguments{
\item{dataset}{dataframe containing numberic columns}

\item{variable}{the variable (e.g. factor) to get unique values from}

\item{order_by_count}{if TRUE (the default) it will plot the bars from most to least frequent, otherwise it will order by the original factor levels if applicable}

\item{base_size}{uses ggplot's base_size parameter for controling the size of the text}
}
\description{
returns a barchart of the unique value counts for a given dataset/variable
}
\examples{

library(ggplot2)
rt_explore_plot_unique_values(dataset=iris, variable='Species')

}
