% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_unique_values}
\alias{rt_explore_unique_values}
\title{returns a count of the unique values for a given dataset/variable}
\usage{
rt_explore_unique_values(dataset, variable)
}
\arguments{
\item{dataset}{dataframe containing numberic columns}

\item{variable}{the variable (e.g. factor) to get unique values from}
}
\description{
returns a count of the unique values for a given dataset/variable
}
\examples{

library(ggplot2)
rt_explore_unique_values(dataset=iris, variable='Species')

}
