% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_floor_date_factor}
\alias{rt_floor_date_factor}
\title{returns an ordered factor of the date vector}
\usage{
rt_floor_date_factor(date_vector, date_floor = "week", week_start = 1)
}
\arguments{
\item{date_vector}{the date vector}

\item{date_floor}{`week`, `month`, `quarter`, `year`}

\item{week_start}{same values of floor_date, defaults to `1` which starts the week on Monday}
}
\description{
date_floor: `week` gives `YYYY-MM-DD` with the `Monday` as the date when week_start is `1`
date_floor: `month` gives `YYYY-MM`
date_floor: `quarter` gives `YYYY-QX`
date_floor: `year` gives `YYYY`
}
