% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rt_geometric_mean}
\alias{rt_geometric_mean}
\title{Calculates the geometric mean of a vector of numbers}
\usage{
rt_geometric_mean(values, na.rm = TRUE, add_subtract = 1e-04)
}
\arguments{
\item{values}{vector of numeric values}

\item{na.rm}{if TRUE then remove NA values, if FALSE then return NA if any NA values exist}

\item{add_subtract}{since the geometric mean takes the log of the values, a value/log of 0 will cause
problems; `add_subtract` adds the number to the values and then subtracts the value after taking the
mean (but before taking the exponent)}
}
\description{
Calculates the geometric mean of a vector of numbers
}
