% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{rt_get_any_touch_attribution}
\alias{rt_get_any_touch_attribution}
\title{gives each step credit for the number of conversions that resulted from the corresponding path conversions}
\usage{
rt_get_any_touch_attribution(.campaign_data, .conversion_column,
  .path_id = ".path_id", .step = "step")
}
\arguments{
\item{.campaign_data}{dataframe with columns `id | timestamp | step | num_conversions | conversion_value`}

\item{.conversion_column}{e.g. num_conversions or conversion_value}

\item{.path_id}{string identifying the path_id column}

\item{.step}{string identifying the step column}
}
\description{
if the path is Facebook -> Facebook -> Facebook -> Instagram -> 2 Conversions, then both Facebook & Instagram get credit for two conversions, regardless how many times they are in the path
}
\details{
returns a dataframe with the combined results
}
