% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_get_cohorted_conversion_rates}
\alias{rt_get_cohorted_conversion_rates}
\title{shows conversion rates (y-axis) of a particular cohort based on various 'snapshots' in time (a line is a snapshot in time).}
\usage{
rt_get_cohorted_conversion_rates(dataset, first_date, second_date,
  reference_date, group_variable = NULL, snapshots = c(1, 7, 14),
  snapshot_units = "days", date_floor = "month")
}
\arguments{
\item{dataset}{dataframe}

\item{first_date}{the reference date (e.g. first-touch point)}

\item{second_date}{the date of conversion, NA if not converted}

\item{reference_date}{we need to know how old the cohort is so we can determine if the}

\item{group_variable}{a variable to group/facet by. If `color_or_facet` will be ignored}

\item{snapshots}{the numeric snapshots}

\item{snapshot_units}{the units of the snapshots e.g. `hours`, `days`, `weeks`}

\item{date_floor}{how we should define the cohort group e.g. by `day`, `by week`, by `months`}
}
\description{
Each record in the cohort must have had enough time to convert, in order to be shown in the graph.
For example, say the snapshots we are interested in are at day 1 and day 7. If the cohort refers to
 February, and it is March 3rd, every record in the February cohort has had at least 1 day of "activity" or
 potential to convert, so we can include Feb for that snapshot. However, not everyone in Feb has had 7 days
 of activity, so we cannot yet include Feb for the 7 day snapshot, because people who joined the cohort on
 Feb 28, for example, have not had a full 7 days of potential activity.
}
