% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_get_colors_from_values}
\alias{rt_get_colors_from_values}
\title{Returns the colors based on the alphabetical order of the values. E.g. if you plot a graph alphabetically
and then plot it based on it's factor order the colors should be retained for each category.}
\usage{
rt_get_colors_from_values(x)
}
\arguments{
\item{x}{a character or factor vector}
}
\description{
Returns the colors based on the alphabetical order of the values. E.g. if you plot a graph alphabetically
and then plot it based on it's factor order the colors should be retained for each category.
}
