% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_get_date_fields}
\alias{rt_get_date_fields}
\title{returns whether or not a value is NULL, NA, or NAN}
\usage{
rt_get_date_fields(date_vector, reference_date = NULL)
}
\arguments{
\item{date_vector}{a vector of dates}

\item{reference_date}{the reference/base date used for fields such as (is_current_xxx). If NULL (default), then today's date is used.}
}
\description{
returns whether or not a value is NULL, NA, or NAN
}
\examples{

date_vector <- lubridate::as_date('2018-01-01') + seq(0, 400)
rt_get_date_fields(date_vector)

}
