% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_get_vector}
\alias{rt_get_vector}
\title{returns a dataframe's column as a vector}
\usage{
rt_get_vector(df, column, return_unique = FALSE)
}
\arguments{
\item{df}{a data.frame}

\item{column}{the column to return as a vector}

\item{return_unique}{if TRUE, return unique values}
}
\description{
returns a dataframe's column as a vector
}
