% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_get_year_month_factors}
\alias{rt_get_year_month_factors}
\title{returns year and month factors in list}
\usage{
rt_get_year_month_factors(.date_vector, .abbreviate = FALSE)
}
\arguments{
\item{.date_vector}{date vector}

\item{.abbreviate}{use month abbreviations}
}
\description{
returns year and month factors in list
}
