% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_group_by_all_of}
\alias{rt_group_by_all_of}
\title{wrapper for group_by for string variables}
\usage{
rt_group_by_all_of(.data, ...)
}
\arguments{
\item{.data}{the data.frame to select from}

\item{...}{string values and/or string vectors}
}
\description{
wrapper for group_by for string variables
}
