% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_is_null_na_nan}
\alias{rt_is_null_na_nan}
\title{returns whether or not a value is NULL, NA, or NAN}
\usage{
rt_is_null_na_nan(x)
}
\arguments{
\item{x}{a single value to check}
}
\description{
returns whether or not a value is NULL, NA, or NAN
}
\examples{

rt_is_null_na_nan(NA)

}
