% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{rt_markov_model}
\alias{rt_markov_model}
\title{wrapper around ChannelAttribution::markov_model}
\usage{
rt_markov_model(
  .path_data,
  .path_sequence = "path_sequence",
  .num_conversions = "num_conversions",
  .conversion_value = "conversion_value",
  .null_conversions = "null_conversions",
  .order = 1,
  .symbol = ">",
  .seed = 42
)
}
\arguments{
\item{.path_data}{dataframe expected by ChannelAttribution::markov_model}

\item{.path_sequence}{var_path}

\item{.num_conversions}{var_conv}

\item{.conversion_value}{var_value}

\item{.null_conversions}{var_null}

\item{.order}{order}

\item{.symbol}{sep}

\item{.seed}{seed}
}
\description{
wrapper around ChannelAttribution::markov_model
}
