% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_params_to_vector}
\alias{rt_params_to_vector}
\title{converts multi-argument values into flattened  vector}
\usage{
rt_params_to_vector(...)
}
\arguments{
\item{...}{values and/or vectors}
}
\description{
converts multi-argument values into flattened  vector
}
