% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_peak}
\alias{rt_peak}
\title{like `head()` but converts to a data.frame so rows/columns aren't excluded when showing}
\usage{
rt_peak(.df, .n = 25)
}
\arguments{
\item{.df}{dataframe}

\item{.n}{number of rows to include}
}
\description{
like `head()` but converts to a data.frame so rows/columns aren't excluded when showing
}
