% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rt_plot_2_proportions_test}
\alias{rt_plot_2_proportions_test}
\title{plots the proportions with confidence intervals according to prop.test}
\usage{
rt_plot_2_proportions_test(
  prop_1,
  prop_2,
  categories,
  groups = NULL,
  confidence_level = 0.95,
  show_confidence_values = TRUE,
  axes_flip = FALSE,
  axis_limits = NULL,
  text_size = 4,
  line_size = 0.35,
  base_size = 11,
  x_label = "",
  y_label = "",
  group_name = "",
  title = "",
  subtitle = NULL,
  caption = NULL
)
}
\arguments{
\item{prop_1}{vector with two values c(numerator, denominator)}

\item{prop_2}{vector with two values c(numerator, denominator)}

\item{categories}{categories}

\item{groups}{vector of groups/categories to plot, seperated by color}

\item{confidence_level}{the confidence level (e.g. 0.95) passed to prop.test}

\item{show_confidence_values}{show the high/low confidence values}

\item{axes_flip}{flip axes}

\item{axis_limits}{custom limits for the corresponding axis (x if not axes_flip, y if axes_flip)}

\item{text_size}{text size (proportion value)}

\item{line_size}{the line size for the error bars}

\item{base_size}{uses ggplot's base_size parameter for controling the size of the text}

\item{x_label}{label for x-axis}

\item{y_label}{label for y-axis}

\item{group_name}{when using `groups`, used for the legend in the plot}

\item{title}{overwrites default title of "Confidence Intervals of Two Proportions"}

\item{subtitle}{subtitle}

\item{caption}{expands on default caption giving p-value}
}
\description{
plots the proportions with confidence intervals according to prop.test
}
