% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{rt_plot_channel_attribution}
\alias{rt_plot_channel_attribution}
\title{wrapper around ChannelAttribution::markov_model & ChannelAttribution::heuristic_models}
\usage{
rt_plot_channel_attribution(
  .channel_attribution,
  .channel_categories = NULL,
  .show_values = TRUE
)
}
\arguments{
\item{.channel_attribution}{dataframe returned by rt_get_channel_attribution()}

\item{.channel_categories}{if provided, colors removal effects by channel categories; named vector with categories as value and channel names as vector names}

\item{.show_values}{show the attribution values}
}
\description{
returns a dataframe with the combined results
}
