% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_plot_colors}
\alias{rt_plot_colors}
\title{Plots the colors from `rt_colors()`}
\usage{
rt_plot_colors(color_names = NULL, sets = NULL)
}
\arguments{
\item{color_names}{filter by the names of the colors}

\item{sets}{filter by the set index}
}
\description{
Plots the colors from `rt_colors()`
}
