% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{rt_plot_markov_removal_effects}
\alias{rt_plot_markov_removal_effects}
\title{graphs removal effects returned by rt_markov_model (i.e. ChannelAttribution::markov_model)}
\usage{
rt_plot_markov_removal_effects(.markov_attribution, .channel_categories = NULL)
}
\arguments{
\item{.markov_attribution}{results from rt_markov_model}

\item{.channel_categories}{if provided, colors removal effects by channel categories; named vector with categories as value and channel names as vector names}
}
\description{
graphs removal effects returned by rt_markov_model (i.e. ChannelAttribution::markov_model)
}
