% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rt_plot_regression_variance_explained}
\alias{rt_plot_regression_variance_explained}
\title{Plot the percent of variance explained for a given regression model}
\usage{
rt_plot_regression_variance_explained(.regression_results)
}
\arguments{
\item{.regression_results}{model}
}
\description{
Plot the percent of variance explained for a given regression model
}
