% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{rt_plot_sankey}
\alias{rt_plot_sankey}
\title{TBD}
\usage{
rt_plot_sankey(
  .path_data,
  .id = "entity_id",
  .path_column = "touch_category",
  .visit_index = "touch_index",
  .global_path_values = NULL,
  .ending_events = NULL,
  .ending_event_fill_name = "End of Data",
  .order_by = c("size", "optimize", "both"),
  .depth_threshold = 4,
  .proportion_threshold_other_category = 0.01
)
}
\arguments{
\item{.path_data}{description}

\item{.id}{description}

\item{.path_column}{description}

\item{.visit_index}{description}

\item{.global_path_values}{description}

\item{.ending_events}{description}

\item{.ending_event_fill_name}{description}

\item{.order_by}{description}

\item{.depth_threshold}{description}

\item{.proportion_threshold_other_category}{description}
}
\description{
TBD
}
