% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{rt_pretty_axes}
\alias{rt_pretty_axes}
\title{format for axes}
\usage{
rt_pretty_axes(x, increase_precision_delta = 0)
}
\arguments{
\item{x}{numeric values}

\item{increase_precision_delta}{number of decimal places to increase or decrease from standard implementation}
}
\description{
format for axes
}
