% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{rt_pretty_dataset}
\alias{rt_pretty_dataset}
\title{Changes string column values and column names to "pretty" strings. Ignores other datatypes else.}
\usage{
rt_pretty_dataset(dataset)
}
\arguments{
\item{dataset}{a dataframe to make pretty}
}
\description{
Changes string column values and column names to "pretty" strings. Ignores other datatypes else.
}
\examples{

library(ggplot2)
rt_pretty_text(c('abc', 'ABC', 'abc/xyz'))

}
