% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_pretty_date_label}
\alias{rt_pretty_date_label}
\title{formats a date}
\usage{
rt_pretty_date_label(date_floor, date_break_format)
}
\arguments{
\item{date_floor}{date floor}

\item{date_break_format}{format}
}
\description{
formats a date
}
