% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{rt_pretty_numbers_long}
\alias{rt_pretty_numbers_long}
\title{Formats numeric values}
\usage{
rt_pretty_numbers_long(values)
}
\arguments{
\item{values}{a numeric vector}
}
\description{
Formats numeric values
}
