% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{rt_pretty_numbers_short}
\alias{rt_pretty_numbers_short}
\title{Formats numeric values}
\usage{
rt_pretty_numbers_short(values)
}
\arguments{
\item{values}{a numeric vector}
}
\description{
Formats numeric values
}
\examples{

rt_pretty_numbers_short(rnorm(n=10, mean=0, sd=0.001))
rt_pretty_numbers_short(rnorm(n=10, mean=1000000, sd=100000))

}
