% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{rt_pretty_percent}
\alias{rt_pretty_percent}
\title{formats a percent}
\usage{
rt_pretty_percent(values, use_na = TRUE)
}
\arguments{
\item{values}{numeric values}

\item{use_na}{if TRUE keep NAs, else return "NA"}
}
\description{
formats a percent
}
