% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{rt_pretty_text}
\alias{rt_pretty_text}
\title{Changes strings to "pretty" strings. Ignores other datatypes else.}
\usage{
rt_pretty_text(values)
}
\arguments{
\item{values}{a string vector}
}
\description{
Changes strings to "pretty" strings. Ignores other datatypes else.
}
\examples{

library(ggplot2)
rt_pretty_text(c('abc', 'ABC', 'abc/xyz'))

}
