% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{rt_prettyNum}
\alias{rt_prettyNum}
\title{wrapper for base `prettyNum`}
\usage{
rt_prettyNum(values, use_na = TRUE)
}
\arguments{
\item{values}{a numeric vector}

\item{use_na}{if TRUE keep NAs, else return "NA"}
}
\description{
wrapper for base `prettyNum`
}
