% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rt_regression_get_ind_var_options}
\alias{rt_regression_get_ind_var_options}
\title{Residual vs. Predicted plot}
\usage{
rt_regression_get_ind_var_options(model, original_dataset,
  dependent_variable)
}
\arguments{
\item{model}{model}

\item{dependent_variable}{}
}
\description{
Residual vs. Predicted plot
}
