% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rt_regression_plot_actual_vs_predicted}
\alias{rt_regression_plot_actual_vs_predicted}
\title{Actual vs. Predicted plot}
\usage{
rt_regression_plot_actual_vs_predicted(model)
}
\arguments{
\item{model}{model}
}
\description{
Actual vs. Predicted plot
}
