% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rt_regression_plot_residual_vs_variable}
\alias{rt_regression_plot_residual_vs_variable}
\title{Residual vs. Predicted plot}
\usage{
rt_regression_plot_residual_vs_variable(model, predictor, dataset)
}
\arguments{
\item{model}{model}

\item{predictor}{the predictor i.e. variable to use}

\item{dataset}{the original dataset}
}
\description{
Residual vs. Predicted plot
}
