% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_remove_val}
\alias{rt_remove_val}
\title{returns the vector (`vec`) without the specified value (`val`). If `val` doesn't exist in `vec`, `vec` is
returned unchanged.}
\usage{
rt_remove_val(vec, val)
}
\arguments{
\item{vec}{the vector}

\item{val}{the value to remove}
}
\description{
returns the vector (`vec`) without the specified value (`val`). If `val` doesn't exist in `vec`, `vec` is
returned unchanged.
}
