% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_select_all_of}
\alias{rt_select_all_of}
\title{wrapper for change in DPLYR where we cannot send vector of strings as variable}
\usage{
rt_select_all_of(.data, ...)
}
\arguments{
\item{.data}{the data.frame to select from}

\item{...}{string values and/or string vectors}
}
\description{
wrapper for change in DPLYR where we cannot send vector of strings as variable
}
