% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{rt_setsubset}
\alias{rt_setsubset}
\title{returns TRUE if `set1` is a subset of `set2`. Returns `FALSE` if either `set1` or `set2` is `NULL` (even
if they are both NULL). The reason for this is because it is used to test sets, and more often then not,
NULL values when testing sets is indicative of a bug.}
\usage{
rt_setsubset(set1, set2)
}
\arguments{
\item{set1}{the set that is a subset}

\item{set2}{the set that is the superset}
}
\description{
returns TRUE if `set1` is a subset of `set2`. Returns `FALSE` if either `set1` or `set2` is `NULL` (even
if they are both NULL). The reason for this is because it is used to test sets, and more often then not,
NULL values when testing sets is indicative of a bug.
}
