% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rt_stopif}
\alias{rt_stopif}
\title{Like `stopifnot`, but stop `stopifnot` stops if the expression is not true, and `rt_stopif` stops if the
expression is true. Avoids the unintuitive double-negative e.g. (`stopifnot(!espression)`) which becomes
`rt_stopif(espression)`.}
\usage{
rt_stopif(exprs)
}
\arguments{
\item{exprs}{the expression}
}
\description{
Like `stopifnot`, but stop `stopifnot` stops if the expression is not true, and `rt_stopif` stops if the
expression is true. Avoids the unintuitive double-negative e.g. (`stopifnot(!espression)`) which becomes
`rt_stopif(espression)`.
}
