% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{rt_str_collapse}
\alias{rt_str_collapse}
\title{See example}
\usage{
rt_str_collapse(.x, .surround, .separate)
}
\arguments{
\item{values}{a string vector}
}
\description{
See example
}
\examples{

rt_str_collapse("example")
rt_str_collapse(c("example", "example2"))

}
