% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times_series.R
\name{rt_ts_auto_regression}
\alias{rt_ts_auto_regression}
\title{adds lag variables, performs regrssion & forecasting, produces a plot}
\usage{
rt_ts_auto_regression(
  dataset,
  dependent_variable = NULL,
  independent_variables = NULL,
  ignore_last_x_periods = NULL,
  lambda = NULL,
  num_lags = NULL,
  include_dependent_variable_lag = TRUE,
  ex_ante_forecast_horizon = NULL,
  build_graphs = TRUE,
  show_dataset_labels = FALSE,
  show_forecast_labels = TRUE
)
}
\arguments{
\item{dataset}{a time-series dataset (single- or multi-variable)}

\item{dependent_variable}{the dependent variable; can be null for single-var datasets}

\item{independent_variables}{the independent variables;

`independent_variables` can be null for single-var datasets; for multi-variable datasets, a value of `NULL` will include all variables

single-variable & multi-variable datasets can also include `trend` and/or `season` values}

\item{ignore_last_x_periods}{exclude last x periods from model}

\item{lambda}{(simply passes value to `tslm`); according to their docs: "Box-Cox transformation parameter. If lambda="auto", then a transformation is automatically selected using BoxCox.lambda. The transformation is ignored if NULL. Otherwise, data transformed before model is estimated."}

\item{num_lags}{if specified, adds lag variables for all the independent_variables}

\item{include_dependent_variable_lag}{if num_lags >= 1 *and* a multi-variable dataset, then this indicates whether or not we should lag the dependent_variable and include it in the model (if the data is single-variable, then this is ignored, because the only thing to lag is the dependent_variable)}

\item{ex_ante_forecast_horizon}{if specified, indicates how far into the future we should forecast. All original variables (except for `trend`/`season`) and all lag values that have a lag less than `ex_ante_forecast_horizon` will be removed.}

\item{build_graphs}{if TRUE, will build various graphs and return those ggplot objects in the returned list}

\item{show_dataset_labels}{whether or not to show the values of the original dataaset}

\item{show_forecast_labels}{whether or not to show the forecast values}
}
\description{
adds lag variables, performs regrssion & forecasting, produces a plot
}
