% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times_series.R
\name{rt_ts_create_lagged_dataset}
\alias{rt_ts_create_lagged_dataset}
\title{returns a new dataset containing the lagged values for specified variables}
\usage{
rt_ts_create_lagged_dataset(
  dataset,
  num_lags = 1,
  lag_variables = NULL,
  keep_variables = NULL
)
}
\arguments{
\item{dataset}{a time-series dataset (single- or multi-variable)}

\item{num_lags}{the number of lags to include (e.g. `3` will include 3 lagged variables for each applicable variable, `x_lag_1`, `x_lag_2`, `x_lag_3`)}

\item{lag_variables}{the variables (i.e. columns) to include lagged values for (the original column will be retained). A value of `NULL` is used for single-variable datasets (i.e. ts datasets that have no columns) or, for multi-variable datasets, will create lagged variables for *all* variables in the dataset.

All variables that are not specified in `lag_variables` are removed.}

\item{keep_variables}{the variables to retain in the dataset which are not specified in `lag_variables`. A value of `NULL` is used for single-variable datasets.}
}
\description{
returns a new dataset containing the lagged values for specified variables
}
