% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times_series.R
\name{rt_ts_get_friendly_time_ticks}
\alias{rt_ts_get_friendly_time_ticks}
\title{Returns friendly axis ticks. e.g. for monthly it returns a factor with values c("2019-01", "2019-02").
If the frequency of the dataset is anything other than 52, 12, 4, it returns the raw time() values.}
\usage{
rt_ts_get_friendly_time_ticks(dataset)
}
\arguments{
\item{dataset}{ts dataset}
}
\description{
Returns friendly axis ticks. e.g. for monthly it returns a factor with values c("2019-01", "2019-02").
If the frequency of the dataset is anything other than 52, 12, 4, it returns the raw time() values.
}
