% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times_series.R
\name{rt_ts_get_time_period}
\alias{rt_ts_get_time_period}
\title{rather than e.g. 2007.333 for May of 2007 that `time()` returns, this returns `2007.01`}
\usage{
rt_ts_get_time_period(ts_values)
}
\arguments{
\item{ts_values}{a time series dataset

library(fpp2)
get_time_period(a10)}
}
\description{
rather than e.g. 2007.333 for May of 2007 that `time()` returns, this returns `2007.01`
}
