% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times_series.R
\name{rt_ts_is_multi_variable}
\alias{rt_ts_is_multi_variable}
\title{determines if a time-series dataset contains a single variable (as opposed to multiple variables)}
\usage{
rt_ts_is_multi_variable(x)
}
\arguments{
\item{x}{a time series dataset

library(fpp2)
rt_ts_is_multi_variable(melsyd)}
}
\description{
determines if a time-series dataset contains a single variable (as opposed to multiple variables)
}
