% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times_series.R
\name{rt_ts_plot_time_series}
\alias{rt_ts_plot_time_series}
\title{plots a time-series dataset}
\usage{
rt_ts_plot_time_series(dataset, show_values = FALSE,
  show_points = FALSE, show_dates = FALSE, y_zoom_min = NA,
  y_zoom_max = NA, facet_multi_variables = FALSE, base_size = 11)
}
\arguments{
\item{dataset}{ts dataset}

\item{show_values}{show text value above each point in time}

\item{show_points}{show points}

\item{show_dates}{show dates above each point in time}

\item{y_zoom_min}{adjust (i.e. zoom in) to the y-axis; sets the minimum y-value for the adjustment}

\item{y_zoom_max}{adjust (i.e. zoom in) to the y-axis; sets the maximum y-value for the adjustment}

\item{facet_multi_variables}{if TRUE each variable gets it's own section}

\item{base_size}{sets the base/text size}
}
\description{
plots a time-series dataset
}
