% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools-package.R
\docType{package}
\name{rtools-package}
\alias{rtools}
\alias{rtools-package}
\title{rtools: Utility Tools for the CoastalWarming Research Team (CIBIO-BIOPOLIS)}
\description{
Utility tools to read EnvLogger data and perform other useful tasks. New tools added regularly.
}
\author{
\strong{Maintainer}: Rui Seabra \email{ptruimar@gmail.com} (\href{https://orcid.org/0000-0002-0240-3992}{ORCID})

Authors:
\itemize{
  \item Fernando Lima \email{fplima@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})
  \item Rita Silva \email{a.ritasilva.15@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})
}

}
\keyword{internal}
