% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{rtools}
\alias{rtools}
\alias{rtools-package}
\title{Some useful methods that I often need.}
\description{
Some useful methods that I often need.
}
\author{
Alexander Lehmann <lehmanna@in.tum.de>
}

