% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envlogger.R
\name{rtools_example}
\alias{rtools_example}
\title{Get paths to example files}
\usage{
rtools_example(pattern = NULL)
}
\arguments{
\item{pattern}{Pattern to select one or more example files. Pattern is vectorized, so more than one value can be supplied. If NULL, all example files are listed.}
}
\value{
The full path to one or more example files, or the filen ames of all example files available.
}
\description{
\code{rtools-package} comes bundled with several sample files in its inst/extdata directory. This function make them easy to access
}
\examples{
# Get the file names of all example files
rtools_example()

# Get the full path to a target example file
rtools_example("04FD_1700_8A6C_05-20230830_180125")

# Get the full paths to any example files matching a search string
rtools_example("2023-08-30_uksca")

# 'rtools_example()' is vectorized, meaning that multiple search strings can be used
rtools_example(c("04E9_6B00_943C_0C", "04CA_A900_E31C_0F"))
}
