% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{Rx}
\alias{Rx}
\title{Homogeneous rotation matrix (x-axis).}
\usage{
Rx(alpha)
}
\arguments{
\item{alpha}{the rotation angle in radians}
}
\value{
a 4x4 homogeneous matrix
}
\description{
A homogeneous 4x4 matrix for rotation around the x-axis.
}

