% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{Ry}
\alias{Ry}
\title{Homogeneous rotation matrix (y-axis).}
\usage{
Ry(alpha)
}
\arguments{
\item{alpha}{the rotation angle in radians}
}
\value{
a 4x4 homogeneous matrix
}
\description{
A homogeneous 4x4 matrix for rotation around the y-axis.
}

