% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{Rz}
\alias{Rz}
\title{Homogeneous rotation matrix (z-axis).}
\usage{
Rz(alpha)
}
\arguments{
\item{alpha}{the rotation angle in radians}
}
\value{
a 4x4 homogeneous matrix
}
\description{
A homogeneous 4x4 matrix for rotation around the z-axis.
}

