% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fun.R
\name{scale_fun}
\alias{scale_fun}
\title{Format axis labels}
\usage{
scale_fun(x)
}
\arguments{
\item{x}{...}
}
\description{
This function formats the labels on your \code{ggplot2} axis labels
}
\examples{
library(ggplot2)
ggplot(iris, aes(x = Sepal.Length, y = Petal.Width, col = Species)) +
geom_point() +
scale_x_continuous(labels = scaleFUN)
}
\author{
Jeppe Vierø
}
