% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptName.R
\name{scriptName}
\alias{scriptName}
\title{The name of the current R script.}
\usage{
scriptName()
}
\value{
a character vector containing the name of the current R script, or
 NA if no --file= argument can be found
}
\description{
The name of the current R script as taken from R's command-line arguments,
e.g. when the script is run by Rscript. More specifically, the name of the
script is extracted from the --file= argument.
}
\seealso{
\code{\link{commandArgs}}
}

