% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{shuffleColumns}
\alias{shuffleColumns}
\title{A shuffled copy of the given dataset \code{X}.}
\usage{
shuffleColumns(X)
}
\arguments{
\item{X}{matrix or data.frame whose columns are supposed to be shuffled}
}
\value{
a shuffled copy of X
}
\description{
A copy of the given dataset whose columns have been shuffled in random order.
}

