% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{shuffleRows}
\alias{shuffleRows}
\title{A shuffled copy of the given dataset \code{X}.}
\usage{
shuffleRows(X)
}
\arguments{
\item{X}{a matrix or data.frame whose rows are supposed to be shuffled}
}
\value{
a shuffled copy of X
}
\description{
A copy of the given dataset whose rows have been shuffled in random order.
}

