% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{stop_if}
\alias{stop_if}
\title{If any of the expressions are TRUE, `stop` is called}
\usage{
stop_if(.expression)
}
\arguments{
\item{.expression}{any number of R expressions which should each evaluatoe to a logical vector}
}
\description{
If any of the expressions are TRUE, `stop` is called
}
