% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{stop_if_any_duplicated}
\alias{stop_if_any_duplicated}
\title{wrapper arround `stop_if(any(duplicated(.x)))`}
\usage{
stop_if_any_duplicated(.x)
}
\arguments{
\item{.x}{vector of e.g. characters}
}
\description{
wrapper arround `stop_if(any(duplicated(.x)))`
}
