% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{stop_if_any_missing}
\alias{stop_if_any_missing}
\title{Wrapper around `stop_if(any_missing(.x, .empty_string_as_missing))`}
\usage{
stop_if_any_missing(.x, .empty_string_as_missing = FALSE)
}
\arguments{
\item{.x}{vector of e.g. characters}

\item{.empty_string_as_missing}{if TRUE, empty strings are treated as missing}
}
\description{
Wrapper around `stop_if(any_missing(.x, .empty_string_as_missing))`
}
