% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{stop_if_not_identical}
\alias{stop_if_not_identical}
\title{Wrapper around `stopifnot(identical(.x, .y))`}
\usage{
stop_if_not_identical(.x, .y)
}
\arguments{
\item{.x}{vector of e.g. characters}

\item{.y}{vector of e.g. characters}
}
\description{
Mainly, this is here so that I remember not to do `stopifnot(all(.x == .y))` because if one of .x or .y has
zero length all() still evaluates to TRUE and the program is not stopped, even though .x and .y do not have
the same values
}
