% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_database.R
\name{subfolders}
\alias{subfolders}
\title{Rearrange the content of an EnvloggerViewer folder into subfolders}
\usage{
subfolders(path, min_size = 700)
}
\arguments{
\item{path}{The path to an EnvLoggerViewer daily folder}

\item{min_size}{Integer indicating the file size below which report files are rejected into a "check" folder (defaults to 700, which is just a few bytes greater than the size of a report with header but no data)}
}
\value{
A tibble with one row for each unique serial name and 5 columns with relevant metadata.
}
\description{
Given a path to a folder from EnvloggerViewer, respective to only one day, reorganize the files into different subfolders by shore name.
}
\examples{
path <- "~/Library/CloudStorage/GoogleDrive-cctbonproject@gmail.com/My Drive/coastalwarming_29_rui/envlogger/2024-05-12"
subfolders(path)
}
\seealso{
\code{\link[=read_env_log]{read_env_log()}}
}
