% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substr_right.R
\name{substr_right}
\alias{substr_right}
\title{Cut character vector from the right}
\usage{
substr_right(x, n)
}
\arguments{
\item{x}{a vector of type \code{character}}

\item{n}{a defined length}
}
\description{
This function allows you to crop a vector of type \code{character} from the right
}
\examples{
substrRight("character vector", 6)
}
\author{
Jeppe Vierø
}
