% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tides.R
\name{tides}
\alias{tides}
\title{Get tide heights for any location (all computations in UTC)}
\usage{
tides(
  lat,
  lon,
  first_day = Sys.Date(),
  last_day = Sys.Date() + 7,
  freq_mins = 10,
  tz = 0,
  mode = "all"
)
}
\arguments{
\item{lat}{Latitude (numeric)}

\item{lon}{Longitude (numeric)}

\item{first_day}{A Date or a string that can be converted with as.Date}

\item{last_day}{A Date or a string that can be converted with as.Date}

\item{freq_mins}{Get tide heights every x mins (numeric)}

\item{tz}{TZ offset (in hours; defaults to zero, which returns times in UTC0)}

\item{mode}{One of "lo" (low tides), "hi" (high tides), "lohi" or "hilo" (low and high tides) or "all" (tide height every \code{freq_mins}) to indicate which tide elements to output}
}
\value{
A tibble with 4 columns:
\itemize{
\item \verb{$time}, timestamps (always UTC + \code{tz})
\item \verb{$h}, tide elevation (in meters) relative to altimeter-derived Mean Water Level
\item \verb{$lo}, logical, \code{TRUE} at low tides
\item \verb{$hi}, logical, \code{TRUE} at high tides
}
}
\description{
Get tide heights for any location (all computations in UTC)
}
\examples{
# Saint Peter Saint Paul Archipelago
(t <- tides(
  lat = 0.917,
  lon = -29.346,
  first_day = "2023-10-14",
  last_day  = "2023-10-19",
  freq_mins = 10
))
plot(t$time, t$h, type = "l")

# Only low tides
# Leixões (Portugal)
(t <- tides(
  lat = 41.1833,
  lon = -8.7000,
  first_day = Sys.Date(),
  last_day  = Sys.Date() + 7,
  freq_mins = 10,
  mode = "lo"
))


}
