% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tides.R
\name{tides_low_stats}
\alias{tides_low_stats}
\title{Get low tide details and stats for several locations (still being worked)}
\usage{
tides_low_stats(x, first_day = Sys.Date(), last_day = Sys.Date() + 7)
}
\arguments{
\item{x}{Either a tibble with 4 columns (loc = location name, lat = latitude, lon = longitude, tz = tz offset) or a path to a csv file that can be read with readr::read_csv to generate such a tibble}

\item{first_day}{A Date or a string that can be converted with as.Date}

\item{last_day}{A Date or a string that can be converted with as.Date}
}
\value{
A tibble with the times and heights for the low tides during the target period, headed by tide amplitude stats
}
\description{
Get low tide details and stats for several locations (still being worked)
}
