% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tides.R
\name{tides_multiple}
\alias{tides_multiple}
\title{Get tide heights for several locations}
\usage{
tides_multiple(
  x,
  first_day = Sys.Date(),
  last_day = Sys.Date() + 7,
  freq_mins = 10,
  mode = "all",
  FES_HANDLER
)
}
\arguments{
\item{x}{Either a tibble with 4 columns (loc = location name, lat = latitude, lon = longitude, tz = tz offset) or a path to a csv file that can be read with readr::read_csv to generate such a tibble}

\item{first_day}{A Date or a string that can be converted with as.Date}

\item{last_day}{A Date or a string that can be converted with as.Date}

\item{freq_mins}{Get tide heights every x mins (numeric)}

\item{mode}{One of "lo" (low tides), "hi" (high tides), "lohi" or "hilo" (low and high tides) or "all" (tide height every \code{freq_mins}) to indicate which tide elements to output}

\item{FES_HANDLER}{A file path pointing to "fes2014/ocean_tide.ini" or a fes_handler}
}
\value{
A tibble with times and tide heights for each target location
}
\description{
Get tide heights for several locations
}
\examples{
x <- tibble::tibble(
    loc = c(
      "Kap Farvel (Greenland)",
      "Scrabster (Scotland)",
      "Sachem Head (USA)",
      "Shinnecock Yacht Club (USA)",
      "Cape Hatteras (USA)",
      "Fernando de Noronha (Brazil)"
      ),
    lat = c(
      59.7500,
      58.6100,
      41.2450,
      40.8183,
      35.2333,
      -3.8333
    ),
    lon = c(
      -43.8833,
      -3.5450,
      -72.7083,
      -72.5533,
      -75.5167,
      -32.4167
    ),
    tz = c(
       -2,
       0,
       -5,
       -5,
       -5,
       -2
    )
 )

tides_multiple(x, FES_HANDLER = "/Users/ruiseabra/RS Dropbox/Rui Seabra/RS/bio/datasets/Rfes2014/src/fes2014/data/fes2014/ocean_tide.ini")
}
