% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardFunctions.R
\name{transf.hist}
\alias{transf.hist}
\title{transf.hist}
\usage{
transf.hist(x, add.value = 0.1)
}
\arguments{
\item{add.value}{The value added to x before transformation (standard = 0.1)}

\item{breaks}{Number of breaks (standard 6)}
}
\description{
Creates different histograms of a variable including non-transformed values
sqrt(x+0.1) and log(x+0.1) transformed values
}
