% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{uncertaintyCoefficient}
\alias{uncertaintyCoefficient}
\title{Uncertainty coefficient of a random variable and corresponding class attribute.}
\usage{
uncertaintyCoefficient(x, y, breaks, base = exp(1), ...)
}
\arguments{
\item{x}{a vector or a factor}

\item{y}{a factor of the same length as x that encodes the class of each
corresponding sample of x}

\item{breaks}{see \code{\link{cut}}}

\item{base}{the base of the logarithm (2 = bits, exp(1) = nats, etc.)}

\item{...}{arguments passed on to \code{\link{entropy}}}
}
\value{
the mutual information of X and Y
}
\description{
\deqn{U(X|Y) = \frac{I(X;Y)}{H(X)}}
}
\seealso{
\code{\link{mutualInformation}} and \code{\link{entropy}}
}

