% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{unitLengthColumns}
\alias{unitLengthColumns}
\title{A copy of \code{X} where each column is a unit vector.}
\usage{
unitLengthColumns(X)
}
\arguments{
\item{X}{a matrix}
}
\value{
a matrix where each column has been scaled to unit length
}
\description{
A copy of \code{X} where each column is a unit vector.
}

