% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{unitLengthRows}
\alias{unitLengthRows}
\title{A copy of \code{X} where each row is a unit vector.}
\usage{
unitLengthRows(X)
}
\arguments{
\item{X}{a matrix}
}
\value{
a matrix where each row has been scaled to unit length
}
\description{
A copy of \code{X} where each row is a unit vector.
}

