% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-doc.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine dataset}
\format{A data frame with 178 observations of 13 variables}
\description{
From the author's page: These data are the results of a chemical analysis of
wines grown in the same region in Italy but derived from three different
cultivars. The analysis determined the quantities of 13 constituents found in
each of the three types of wines.
}
\details{
See also \url{http://archive.ics.uci.edu/ml/datasets/Wine}.
}
\author{
Stefan Aeberhard <stefan@coral.cs.jcu.edu.au>
}
\keyword{datasets}

