% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{wmt_to_merc}
\alias{wmt_to_merc}
\title{Web Map Tiles}
\usage{
wmt_to_merc(x, y, z)
}
\arguments{
\item{x}{A numeric value or vector}
}
\value{
A numeric value or vector
}
\description{
Convert Web Map Tile coordinates to Mercator
}
\examples{
wmt_to_merc(sample(0:15), sample(0:15), 4)
}
