% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardFunctions.R
\name{write.clipboard}
\alias{write.clipboard}
\title{Write to clipboard}
\usage{
write.clipboard(dataset, row.names = FALSE)
}
\arguments{
\item{dataset}{The dataset to write as a table to the clipboard (MAC)}
}
\description{
Standard separator is tab and row.names = FALSE
}
\examples{
write.clipboard (table1)
}
