% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{zip.c}
\alias{zip.c}
\title{Tuples zipped by using the \code{c()} function.}
\usage{
zip.c(...)
}
\arguments{
\item{...}{the lists/vectors/etc. that hold the elements to be zipped}
}
\value{
a list of tuples
}
\description{
Tuples zipped by using the \code{c()} function.
}
\seealso{
\code{\link{zip}}.
}

