% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{zip.list}
\alias{zip.list}
\title{Tuples zipped by using the \code{list()} function.}
\usage{
zip.list(...)
}
\arguments{
\item{...}{the lists/vectors/etc. that hold the elements to be zipped}
}
\value{
a list of tuples
}
\description{
Tuples zipped by using the \code{list()} function.
}
\seealso{
\code{\link{zip}}.
}

