% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{zip}
\alias{zip}
\title{Tuples consisting of the first, second, third, and so forth elements of the
given arguments in \code{...}.}
\usage{
zip(zipper, ...)
}
\arguments{
\item{zipper}{the function used for concatenation, e.g. \code{c} or \code{list}}

\item{...}{the lists/vectors/etc. that hold the elements to be zipped}
}
\value{
a list of tuples
}
\description{
Tuples consisting of the first, second, third, and so forth elements of the
given arguments in \code{...}.
}

