% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_get_param}
\alias{aws_get_param}
\title{AWS tools}
\usage{
aws_get_param(secret_id, region = "eu-west-2")
}
\arguments{
\item{secret_id}{Character string AWS parameter name}

\item{region}{Character string AWS region}
}
\value{
A character string
}
\description{
Retrieve AWS secret
}
\examples{
aws_get_param("param-name")
}
