% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hex.R
\name{get_hex}
\alias{get_hex}
\title{Format axis labels}
\usage{
get_hex(x, y)
}
\arguments{
\item{x}{...}
}
\description{
This function formats the labels on your \code{ggplot2} axis labels
}
\examples{
x <- c("14 17 22")
hex <- sapply(strsplit(x, " "), function(x) rgb(x[1], x[2], x[3], maxColorValue=255))
}
\author{
Jeppe Vierø
}
