% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envlogger.R
\name{read_env_data}
\alias{read_env_data}
\title{Read the data from an EnvLogger file}
\usage{
read_env_data(path, skip, zero_secs = TRUE, check = TRUE)
}
\arguments{
\item{path}{Path to an EnvLogger file}

\item{skip}{A numeric value indicating how many rows to skip when reading the target EnvLogger file to skip the header (the \verb{$skip} column of the output from \code{read_env_header()}).}

\item{zero_secs}{Logical, whether to remove trailing seconds (only used if path points to an EnvLogger file; adjustment determined based on the first timestamp)}

\item{check}{Logical, whether or not to check if path points to an EnvLogger file. Defaults to \code{TRUE}. Only meant to be \code{FALSE} when called from within \code{read_env()}, so as not to repeat the same check multiple times. \strong{Do not set it to \code{FALSE} when running this function directly.}
#'}
}
\value{
A tibble with 2 columns (\code{t} and \code{temp}) header$nrow` rows and 2 columns .
}
\description{
Given a file path, check if it points to an EnvLogger file and, if \code{TRUE}, read just the data
}
\examples{
path <- rtools_example()[3]
read_env_header(path)
}
\seealso{
\code{\link[=read_env_header]{read_env_header()}}, \code{\link[=read_env]{read_env()}}, \code{\link[=plot_env]{plot_env()}}
}
