% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_numeric_summary}
\alias{rt_explore_numeric_summary}
\title{returns a dataframe containing summary statistics for numeric columns passsed to `dataset`}
\usage{
rt_explore_numeric_summary(dataset)
}
\arguments{
\item{dataset}{dataframe containing numeric columns}
}
\description{
returns a dataframe containing summary statistics for numeric columns passsed to `dataset`
}
\examples{

library(ggplot2)
rt_explore_numeric_summary(iris)

}
