% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_plot_categoric_heatmap}
\alias{rt_explore_plot_categoric_heatmap}
\title{returns a heat map with histogram on top/side showing distribution of both features.}
\usage{
rt_explore_plot_categoric_heatmap(dataset, x_variable, y_variable,
  sum_by_variable = NULL, count_distinct_variable = NULL,
  multi_value_delimiter = NULL, rev_na_factor_y = FALSE,
  include_percentages = TRUE, base_size = 11)
}
\arguments{
\item{dataset}{dataframe containing numberic columns}

\item{x_variable}{the categoric variable to show on the x-axis}

\item{y_variable}{the categoric variable to show on the y-axis}

\item{sum_by_variable}{the numeric variable to sum}

\item{count_distinct_variable}{when aggregating, rather than counting the total number of records, count distinct occurances of this variabled (cannot be used with `sum_by_variable`)}

\item{multi_value_delimiter}{(currently not implemented) if the variable contains multiple values (e.g. "A", "A, B", ...) then setting
this variable to the delimiter will cause the function to count seperate values}

\item{rev_na_factor_y}{if TRUE, the <NA> factor is added as the first level (top of graph) rather than last level (bottom of graph)}

\item{include_percentages}{in addition to the total count, also display the percent (out of the total number of records; or total sum_by_variable if used )}

\item{base_size}{uses ggplot's base_size parameter for controling the size of the text}
}
\description{
returns a heat map with histogram on top/side showing distribution of both features.
}
