% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_plot_correlations}
\alias{rt_explore_plot_correlations}
\title{returns a heatmap of correlations}
\usage{
rt_explore_plot_correlations(dataset, corr_threshold = 0,
  p_value_threshold = 1, type = "pearson", base_size = 11)
}
\arguments{
\item{dataset}{dataframe containing numberic columns}

\item{corr_threshold}{any correlations that are <= `corr_threshold` will be set to `NA`. (Default is `0`, so all correlations are shown). Helps to reduce noise.}

\item{p_value_threshold}{any correlations that have a p-value greater than `p_value_threshold` will be set to `NA` (Default is `1`, so all correlations are shown)}

\item{type}{type of correlation to perform (Default is `pearson`)}

\item{base_size}{uses ggplot's base_size parameter for controling the size of the text}
}
\description{
returns a heatmap of correlations
}
\examples{

library(ggplot2)
rt_explore_correlations(iris)

}
