% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_plot_value_totals}
\alias{rt_explore_plot_value_totals}
\title{returns a barchart of the unique value counts for a given dataset/variable, grouped by an additional variable}
\usage{
rt_explore_plot_value_totals(dataset, variable,
  comparison_variable = NULL, sum_by_variable = NULL,
  order_by_count = TRUE, show_variable_totals = TRUE,
  show_comparison_totals = TRUE, stacked_comparison = FALSE,
  base_size = 11)
}
\arguments{
\item{dataset}{dataframe containing numberic columns}

\item{variable}{the variable (e.g. factor) to get unique values from}

\item{comparison_variable}{the additional variable to group by; must be a string/factor column}

\item{sum_by_variable}{the numeric variable to sum}

\item{order_by_count}{if TRUE (the default) it will plot the bars from most to least frequent, otherwise it will order by the original factor levels if applicable}

\item{show_variable_totals}{if TRUE (the default) the graph will display the totals for the variable}

\item{show_comparison_totals}{if TRUE (the default) the graph will display the totals for the comparison_variable}

\item{stacked_comparison}{rather than side-by-side bars for the comparison variable, the bars are stacked within the main variable}

\item{base_size}{uses ggplot's base_size parameter for controling the size of the text}
}
\description{
returns a barchart of the unique value counts for a given dataset/variable, grouped by an additional variable
}
