% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_plot_value_totals}
\alias{rt_explore_plot_value_totals}
\title{returns a barchart of the unique value counts for a given dataset/variable, grouped by an additional variable}
\usage{
rt_explore_plot_value_totals(dataset, variable,
  comparison_variable = NULL, sum_by_variable = NULL,
  order_by_count = TRUE, show_variable_totals = TRUE,
  show_comparison_totals = TRUE, stacked_comparison = FALSE,
  multi_value_delimiter = NULL, base_size = 11)
}
\arguments{
\item{dataset}{dataframe containing numberic columns}

\item{variable}{the variable (e.g. factor) to get unique values from}

\item{comparison_variable}{the additional variable to group by; must be a string/factor column}

\item{sum_by_variable}{the numeric variable to sum}

\item{order_by_count}{if TRUE (the default) it will plot the bars from most to least frequent, otherwise it will order by the original factor levels if applicable}

\item{show_variable_totals}{if TRUE (the default) the graph will display the totals for the variable}

\item{show_comparison_totals}{if TRUE (the default) the graph will display the totals for the comparison_variable}

\item{stacked_comparison}{rather than side-by-side bars for the comparison variable, the bars are stacked within the main variable}

\item{multi_value_delimiter}{if the variable contains multiple values (e.g. "A", "A, B", ...) then setting
this variable to the delimiter will cause the function to count seperate values}

\item{base_size}{uses ggplot's base_size parameter for controling the size of the text}
}
\description{
If `multi_value_delimiter` is not NULL, then it counts all the values found after it splits/separates the
     variable by the delimiter. If `sum_by_variable` is NULL, it counts the values and the denominator for
     the `percent` column returned is the total number of records. If `sum_by_variable` is not NULL, then
     when multiple values are found, each value is weighted by the value found in `sum_by_variable`, and
     the denominator for the `percent` column returned is the `sum` of `sum_by_variable` (before the values
     are split).
}
\details{
Currently only works when using only `variable` (not `comparison_variable`)
}
