% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_explore_value_totals}
\alias{rt_explore_value_totals}
\title{returns either a *count* of the unique values of `variable` if `sum_by_variable` is NULL, otherwise it *sums* the
variable represented by `sum_by_variable` across (i.e. grouped by) `variable`}
\usage{
rt_explore_value_totals(dataset, variable, second_variable = NULL,
  count_distinct = NULL, sum_by_variable = NULL,
  multi_value_delimiter = NULL)
}
\arguments{
\item{dataset}{dataframe containing numberic columns}

\item{variable}{the variable (e.g. factor) to get unique values from}

\item{second_variable}{group by a second variable}

\item{count_distinct}{count the distinct number of values in this column}

\item{sum_by_variable}{the numeric variable to sum}

\item{multi_value_delimiter}{if the variable contains multiple values (e.g. "A", "A, B", ...) then setting
this variable to the delimiter will cause the function to count seperate values}
}
\description{
If `multi_value_delimiter` is not NULL, then it counts all the values found after it splits/separates the
     variable by the delimiter. If `sum_by_variable` is NULL, it counts the values and the denominator for
     the `percent` column returned is the total number of records. If `sum_by_variable` is not NULL, then
     when multiple values are found, each value is weighted by the value found in `sum_by_variable`, and
     the denominator for the `percent` column returned is the `sum` of `sum_by_variable` (before the values
     are split).
}
\examples{

library(ggplot2)
rt_explore_value_totals(dataset=iris, variable='Species')

}
