% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_basics.R
\name{rt_plot_proportions}
\alias{rt_plot_proportions}
\title{plots the proportions with confidence intervals according to prop.test}
\usage{
rt_plot_proportions(numerators, denominators, categories, text_size = 4,
  line_size = 0.35, x_label = "", y_label = "", title = "")
}
\arguments{
\item{numerators}{numerators}

\item{denominators}{denominators}

\item{categories}{categories}

\item{text_size}{text size (proportion value)}

\item{line_size}{the line size for the error bars}

\item{x_label}{label for x-axis}

\item{y_label}{label for y-axis}

\item{title}{title}
}
\description{
plots the proportions with confidence intervals according to prop.test
}
