% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{rt_plot_sankey}
\alias{rt_plot_sankey}
\title{creates a sankey diagram based on a data-frame containing touch-points in the form of}
\usage{
rt_plot_sankey(
  .path_data,
  .id = "entity_id",
  .path_column = "touch_category",
  .visit_index = "touch_index",
  .weight = NULL,
  .valid_final_touch_points = NULL,
  .ensure_complete_funnel = TRUE,
  .bounced_fill_value = "Bounced",
  .no_prior_data = "<No Prior Touch-Point>",
  .global_path_values = NULL,
  .depth_threshold = NULL,
  .order_by = c("size", "optimize", "both")
)
}
\arguments{
\item{.path_data}{data-frame of touch-points}

\item{.id}{entity id column}

\item{.path_column}{column of the path or touch-points}

\item{.visit_index}{numeric index touch-points}

\item{.weight}{column that gives a weight value, which should be the same for each touch-point for a given entity_id (basically the value of the entity)}

\item{.valid_final_touch_points}{specify what is considered the ending events.}

\item{.ensure_complete_funnel}{If TRUE, this makes sure that the number of people/entities in the first level
of the sankey plot equals the number of people/entities in the final level of the plot
Any path where the last touch-point is not one of the values in
`.valid_final_touch_points` is considered a "bounce" and an additional touch-point will be added.
The value of the new touch-point will be based on `.bounced_fill_value`
Any path that only 1 touch-point and the value is in `.valid_final_touch_points` will have an additional
touch-point added of "<No Prior Touch-Point>"}

\item{.bounced_fill_value}{the value to be inserted as a touch-point when an entity bounces (does not have a conversion touch-point)}

\item{.no_prior_data}{the value to be inserted as a touch-point when an entity only has a conversion touch-point}

\item{.global_path_values}{a list of all possible touch-categories, which the colors will be based on. This
ensures that the colors of graphs are consistent across multiple graphs even if the same categories
don't show up in each graph}

\item{.depth_threshold}{the number of leves at the beginning and end of the journey/path}

\item{.order_by}{determines the arrangement of the diagram; `both` returns a list of 2 items, each containing a plot}
}
\description{
entity_id | touch_category | touch_index | weight (optional)
    ============================================================
    personA   | Home Page      | 1             0
    personA   | Pricing Page   | 2             0
    personA   | Conversion     | 3             0
    personB   | Home Page      | 1             2
    personB   | Bounced        | 2             2
    personC   | Pricing        | 1             1
}
\details{
where the column names are specified as parameters.
}
