% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{rt_pretty_numbers_long}
\alias{rt_pretty_numbers_long}
\title{Formats numeric values}
\usage{
rt_pretty_numbers_long(values, increase_precision_delta = 0)
}
\arguments{
\item{values}{a numeric vector}

\item{increase_precision_delta}{number of decimal places to increase or decrease from standard implementation}
}
\description{
Formats numeric values
}
