% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{rt_str_collapse}
\alias{rt_str_collapse}
\title{See example}
\usage{
rt_str_collapse(.x, .surround, .separate)
}
\arguments{
\item{.x}{a vector of string values}

\item{.surround}{string to surround each string in the vector}

\item{.separate}{string to separate each string in the vector}
}
\description{
See example
}
\examples{

rt_str_collapse("example", "'", " + ")
rt_str_collapse(c("example", "example2"), "'", " + ")

}
