% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times_series.R
\name{rt_ts_auto_regression}
\alias{rt_ts_auto_regression}
\title{adds lag variables, performs regrssion & forecasting, produces a plot}
\usage{
rt_ts_auto_regression(dataset, dependent_variable = NULL,
  independent_variables = NULL, num_lags = NULL,
  ex_ante_forecast_horizon = NULL, show_dataset_labels = FALSE,
  show_forecast_labels = TRUE)
}
\arguments{
\item{dataset}{a time-series dataset (single- or multi-variable)}

\item{dependent_variable}{the dependent variable; can be null for single-var datasets}

\item{independent_variables}{the independent variables;

`independent_variables` can be null for single-var datasets; for multi-variable datasets, a value of `NULL` will include all variables

single-variable & multi-variable datasets can also include `trend` and/or `season` values}

\item{num_lags}{if specified, adds lag variables for all the independent_variables}

\item{ex_ante_forecast_horizon}{if specified, indicates how far into the future we should forecast. All original variables (except for `trend`/`season`) and all lag values that have a lag less than `ex_ante_forecast_horizon` will be removed.}

\item{show_dataset_labels}{whether or not to show the values of the original dataaset}

\item{show_forecast_labels}{whether or not to show the forecast values}
}
\description{
adds lag variables, performs regrssion & forecasting, produces a plot
}
